% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-mean.R
\name{area_mean}
\alias{area_mean}
\title{Isobath (200m) separated summary of values across an area}
\usage{
area_mean(x)
}
\arguments{
\item{x}{pacea 'sf' data object with only one column of values (and a geometry column)}
}
\value{
list of summary statistics
}
\description{
Estimate summary statistics across a spatial area, and separated by the 200m isobath that indicates inshore and offshore regions.
}
\examples{
\dontrun{
pdata <- bccm_surface_temperature()

yrs <- c(2019)
mths <- c(8) # august
ym <- paste(yrs, mths, sep = "_")

tdat <- pdata \%>\% dplyr::select(all_of(ym))

area_mean(tdat)
}
}
