% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_coast}
\alias{bc_coast}
\title{BC coastline}
\format{
A simple features multipolygon dataframe.
}
\source{
Generated from running `data-raw/coastline/coastline-eez.R`.
}
\usage{
bc_coast
}
\description{
A simple features object of the BC coastline
}
\details{
Extent of BC coast includes the BC coast and Pacific portion of Canada's Exclusive Economic Zone, and any neighbouring land (e.g. US States). 

Source of initial bc_coast geometry shapes were loaded from 'rnaturalearth::ne_countries' at a 1:10M scale.
}
\examples{
\dontrun{
bc_coast
plot(bc_coast)
}

}
\keyword{datasets}
