% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_eez}
\alias{bc_eez}
\title{BC Exclusive Economic Zone (EEZ)}
\format{
A simple features polygon dataframe.
}
\source{
Generated from running `data-raw/coastline/coastline-eez.R`.
}
\usage{
bc_eez
}
\description{
A simple features object of the BC EEZ
}
\details{
Source of BC EEZ is from 'PBSdata'.
}
\examples{
\dontrun{
bc_eez
plot(bc_eez)
}

}
\keyword{datasets}
