% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_Major_Area_Boundaries.R
\docType{data}
\name{BC_Major_Area_Boundaries}
\alias{BC_Major_Area_Boundaries}
\title{Contains the Definitions of 4 Major Regions: WCHG, WCVI, HS, QCS}
\format{
A SpatialPolygons object:
}
\usage{
BC_Major_Area_Boundaries
}
\description{
A SpatialPolygonsDataFrame object
}
\keyword{datasets}
