% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts-get-data.R
\name{bccm_all_variables}
\alias{bccm_all_variables}
\title{Download all BCCM data}
\usage{
bccm_all_variables()
}
\value{
downloaded files to `pacea_cache()` directory
}
\description{
This function downloads all BCCM model data.
}
\details{
Use `bccm_data` To view the full list of available variables

CAUTION: The BCCM files are large and downloading all data may take a while.
}
\examples{
\dontrun{
bccm_all_variables()
}
}
