% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bccm_data}
\alias{bccm_data}
\title{BCCM data layers}
\format{
table
}
\source{
Pena et al., 2019.
Generated from running `data-raw/data-key/data-list.R`
}
\usage{
bccm_data
}
\description{
List of available BCCM data layers for download.
}
\examples{
\dontrun{
bccm_data
}

}
\keyword{datasets}
