% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_climatology_anomaly.R
\name{calc_anom}
\alias{calc_anom}
\title{Calculate anomaly}
\usage{
calc_anom(
  data,
  clim_years = c(1991:2020),
  clim_time = "month",
  time_period_return = "all",
  years_return
)
}
\arguments{
\item{data}{pacea data object: BCCM, OISST, buoy_sst data only}

\item{clim_years}{climatology period years}

\item{clim_time}{time units (e.g. month) to summarize climatologies}

\item{time_period_return}{vector of value(s) for the specific time units to estimate climatologies (e.g. '4' for week 4 or April). Set to equal 'all' for all time units.}

\item{years_return}{vector of value(s) to return the years of interest. Defaults to all years in input data}
}
\value{
anomaly data values of same class as input pacea data
}
\description{
Function for calculating climatology of a pacea data object.
}
\details{
The functions \code{calc_clim} and \code{calc_anom} output the climatology of a specified time period and anomaly values relative to a climatological period, respectively. These functions can be used for data from BCCM (e.g. \code{bccm_surface_temperature()}), OISST (e.g. \code{oisst_7day}), and buoy SST (e.g. \code{buoy_sst}).
}
\examples{
\dontrun{
pdata <- bccm_surface_temperature()
anom_bccm <- calc_anom(pdata)
head(anom_bccm)

anom_bccm2 <- calc_anom(pdata, time_period_return = c(1, 6), years_return = 1999)
head(anom_bccm2)

# oisst data
anom_oi <- calc_anom(oisst_7day)
head(anom_oi)

# pacea buoy sst data
anom_buoy <- calc_anom(buoy_sst)
head(anom_buoy)

anom_buoy2 <- calc_anom(buot_sst, clim_time = "week")
head(anom_buoy2)
}
}
