% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-anomalies-pacea-oi.R
\name{calculate_anomalies.pacea_oi}
\alias{calculate_anomalies.pacea_oi}
\title{TODO all the help Calculate climatology and anomalies for buoy SST data.}
\usage{
\method{calculate_anomalies}{pacea_oi}(
  data,
  climatology_years = 1991:2020,
  climatology_time = "month",
  time_period_return = "all",
  years_return = NULL,
  area = NULL
)
}
\arguments{
\item{data}{\code{buoy_sst} pacea data object}

\item{climatology_years}{climatology period years}

\item{climatology_time}{time units ("month" or "week") to summarize
climatologies and anomalies, but only "month" is currently available (there
is some functionality to use "week" also -- contact Andy if this would be useful).}

\item{time_period_return}{vector of value(s) for the specific time units to estimate climatologies (e.g. '4' for week 4 or April). Set to equal 'all' for all time units.}

\item{years_return}{vector of value(s) to return the years of
interest. Defaults to all years in input data}

\item{min_days_per_month}{minimum number of daily SST values required in a
time period (month or week) for that period to be included in climatology
calculation and anomalies calculation. Defaults to 15 days per month.}

\item{max_consecutive_missing_days}{maximum number of consecutive days allowed
to be missing (NA) within a time period (month or week). If a time period has
more than this many consecutive NAs, it is excluded from climatology and anomalies
calculation. Defaults to 6 days.}
}
\value{
pacea_buoy_anomalies_list object (also a list) of climatology of data and anomalies, of class
\code{pacea_buoy_anomalies_list}. Note that \code{climatology_years} will be the
prescribed years, but these may not be available for all buoys.
}
\description{
Function for calculating climatology and subsequent anomalies of a \code{buoy_sst} pacea data object.
}
\examples{
\dontrun{
area_126 <- tibble(
  lat = c(-127.1506, -128.2331, -129.3492, -127.9167, -127.1847, -126.8200,
-127.1506),
  lon = c(49.85766, 49.00000, 48.99991, 50.11915, 50.40183, 50.24466, 49.85766))
res <- calculate_anomalies(oisst_month, area = area_126)
}
}
\author{
Andrew Edwards and Travis Tai
}
