% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-index-changed.R
\name{check_index_changed}
\alias{check_index_changed}
\title{Check if a newly downloaded pacea object has changed from one currently in pacea}
\usage{
check_index_changed(old, new, alpi = FALSE)
}
\arguments{
\item{old}{original object (of class \code{pacea_index}, \code{pacea_recruitment},
\code{pacea_biomass}, or \code{pacea_recruitment_herring}) saved in pacea}

\item{new}{newly downloaded and wrangled object (must also be one of the above classes)}

\item{alpi}{logical if checking the \code{alpi} object, as comparing axis name
does not work because it is convoluted. See
\code{data-raw/coastwide-indices/alpi-update.R} for use.}
}
\value{
logical if the object has changed (and we would then want to resave it)
}
\description{
If index has changed from the saved one then save the new version. Only used
for updating data (see files in \code{data-raw}).
}
\examples{
\dontrun{
check_index_changed(oni, oni[-10, ]))
# see data-raw/groundfish/data.R
# see data-raw/coastwide-indices/coastwide-indices.R
}
}
\author{
Andrew Edwards
}
