% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data-hake.R
\name{create_data_hake}
\alias{create_data_hake}
\title{Create package data with one call. Taken from hake package.}
\usage{
create_data_hake(var, val)
}
\arguments{
\item{var}{The variable name}

\item{val}{The value for variable `var`}
}
\value{
Nothing, creates package data. See `usethis::use_data()`.
}
\description{
Usually you have to create a variable first, and then call
 [usethis::use_data()] with the variable name as an argument and the
 package data will be created. This function just unifies that into one step,
 and allows automated naming with `paste0()`, unlike `usethis::use_data()`.
 Keeping hake in the name for consistency with the original function, but this
 will work for any data object. See `data-raw/groundfish/hake.R` for example use.
}
\examples{
\dontrun{
In data-raw/groundfish/hake.R see these lines (which allow `assess_yr` to
  part of the data object, which `use_data()` does not):
 create_data_hake(paste0("hake_recruitment_", assess_yr),
                  paste0("hake_recruitment_", assess_yr))
}
}
\author{
Chris Grandin (help file added to by Andrew Edwards)
}
