% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-zenodo-data.R
\name{get_zenodo_data}
\alias{get_zenodo_data}
\title{Get large data products from Zenodo}
\usage{
get_zenodo_data(
  layer,
  update = FALSE,
  ask = interactive(),
  force = FALSE,
  version = "01",
  cache_subfolder = "hotssea"
)
}
\arguments{
\item{layer}{Name of the data object.}

\item{update}{Logical. Would you like to check for a newer version of the layer?}

\item{ask}{Logical. Should the user be asked before downloading the data to local cache? Defaults to the value of interactive().}

\item{force}{Logical. Should download of data be forced? Overrides `ask`
argument if TRUE. Does not force a download if the file is already cached locally.}

\item{version}{Character. Version number of data on Zenodo (as in what is
appended at the end of the filenames).}

\item{cache_subfolder}{Character. The folder
`paste0(pacea::pacea_cache(), cache_subfolder)` will be
searched first for the desired model output file, and if the no file is
there then it will be downloaded from Zenodo into this directory.}
}
\value{
Data object requested
}
\description{
Assumes version `_01` for now. Have to adapt code if version gets updated,
but it's quite fiddly (though Travis did for BCCM).  TODO
}
\details{
TODO update if get same function working for bccm_full.

HOTSSea model results have been wrangled from the original results provided
 by Greig Oldford (see data-raw/hotssea). These are too big to be included in
 pacea as standard data objects, and so have been archived on Zenodo. The
 function downloads an object and saves it in on the user's computer. It is
 automatically called from TODO, and is not really needed separately by
 users. See that for usage.

Based on [get_bccm_data()], but getting from Zenodo not `pacea-data/`.
}
\examples{
\dontrun{
tt <- get_zenodo_data("hotssea_surface_temperature_max")
plot(tt)
}

}
\author{
Andrew Edwards and Travis Tai
}
