% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacea.R
\docType{package}
\name{gfiphc}
\alias{gfiphc}
\title{pacea package dplyr-related things}
\description{
Adapting from https://github.com/pbs-assess/gfiphc/blob/master/R/gfiphc.R
(see that if get issues with column names within dplyr functions for what to
 add). And from
https://github.com/andrew-edwards/sizeSpectra/blob/master/R/sizeSpectra.R for
some tibble related stuff.
}
