% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid26}
\alias{grid26}
\title{Geospatial pacea grid - BC coastline grid template}
\format{
A simple features dataframe.
}
\source{
Generated from running \code{data-raw/grids/create-grid.R}.
}
\usage{
grid26
}
\description{
A simple features object of the grid used for interpolation of British Columbia continental margin (BCCM) model. The grid is divided into two regions with differing resolutions: a higher resolution inshore 2km x 2km grid and lower resolution offshore 6km x 6km grid. The boundary between the two regions is created using the 200m isobath along the continental shelf.
}
\details{
Grid cells were removed where the overlapped land using the 'bc_coast' shapefile in pacea. The grid only covers the Canadian Pacific EEZ region with a 10km buffer.
}
\examples{
\dontrun{
head(grid26)
plot(grid26)

# convert to sf points object
library(sf)
grid26_points <- st_centroid(grid26)
}

}
\keyword{datasets}
