% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hake_biomass}
\alias{hake_biomass}
\alias{hake_biomass_2023}
\alias{hake_biomass_2024}
\alias{hake_biomass_2025}
\title{Pacific Hake annual spawning stock biomass (females only) as estimated by
the most recent (2025) stock assessment.}
\format{
A tibble also of class \code{pacea_biomass} with columns:
\describe{
\item{year:}{year of the estimate of spawning stock biomass (mature
females), where the estimate is for the start of the year}
\item{low:}{low end (2.5th percentile) of the 95\\% credible interval for
biomass, units as for \code{median}}
\item{median:}{median estimate of biomass, in millions of tonnes of female
spawning biomass at the start of the year}
\item{high:}{high end (97.5th percentile) of the 95\\% credible interval for
biomass, units as for \code{median}}
}

An object of class \code{pacea_biomass} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 58 rows and 4 columns.

An object of class \code{pacea_biomass} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 59 rows and 4 columns.

An object of class \code{pacea_biomass} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 60 rows and 4 columns.
}
\source{
Generated from Andy running (in the hake repository)
\code{pacea_save()} and then here \code{data-raw/groundfish/hake.R}.
}
\usage{
hake_biomass

hake_biomass_2023

hake_biomass_2024

hake_biomass_2025
}
\description{
The Pacific Hake stock is managed and assessed through an Agreement between
Canada and the United States. The spawning stock biomass (mature females) estimates come from the most
recent joint stock assessment, and, importantly, are for the coastwide stock
from California to British Columbia.
}
\details{
See \code{?hake_recruitment} for further details and references, and explanations
of \code{hake_biomass_2025}, \code{hake_biomass_2024}, and \code{hake_biomass_2023}.
}
\examples{
\dontrun{
hake_biomass
plot(hake_biomass)
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
