% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hake_recruitment}
\alias{hake_recruitment}
\title{Pacific Hake annual age-0 recruitments as estimated by the 2023 stock
assessment.}
\format{
A tibble also of class `pacea_recruitment` with columns:
\describe{
  \item{year:}{year of the recruitment estimate}
  \item{low:}{low end (2.5th percentile) of the 95\% credible interval for
  recruitment, billions of age-0 fish}
  \item{median:}{median estimate of recruitment, billions of age-0 fish}
  \item{low:}{high end (97.5th percentile) of the 95\% credible interval for
  recruitment, billions of age-0 fish}
 }
}
\source{
Generated from Andy running (in the hake repository)
  `hake-assessment/sandbox/andy/pacea-save/pacea-save.R` and then here
  `data-raw/groundfish/hake.R`.
}
\usage{
hake_recruitment
}
\description{
The Pacific Hake stock is managed and assessed through an Agreement between
Canada and the United States. The recruitment estimates come from the most
recent joint stock assessment, and, importantly, are for the coastwide stock
from California to British Columbia.
}
\details{
Pacific Hake are a migratory species, generally spawning off southern
California in the winter spawning season and migrating to coastal areas
between northern California and northern British Columbia during the rest of
the year. The stock tends to move farther to the north in the summer during
warmer years compared to cool years.

The stock is important to ecosystem dynamics in the Eastern
Pacific Ocean due to its relatively large total biomass and potentially large
role as both prey and predator. It has highly variable recruitment, as seen
by typing `plot(hake_recruitment)`.

The stock is assessed as a single migratory stock, so it is important to
realise (in the context of analyses for Canadian waters) that these estimates are for the
full coastwide stock. Results from the biannual coastwide acoustic survey show how the
biomass of age-2 and older fish in Canadian waters can vary between years
(Figure 2 of most recent assessment).

The `hake_recruitment` (and `hake_biomass`) estimates are from a Bayesian statistical catch-at-age
model, that is fit to the acoustic survey index of biomass, an index of
age-1 fish from the survey, annual commercial catch-at-age data, and
age-composition data from the survey and commercial fisheries.

Note that recruitments for current and recent years have little data to
estimate them, and so can essentially come from the assumed distribution of
recruitments. They should therefore likely be excluded from any analyses, as
they are not based on data. Exactly how many years are affected can depend
upon the timing of the assessment relative to the most recent biannual
survey.

Results are from the most recent stock assessment:

Berger, A.M., C.J. Grandin, K.F. Johnson and A.M. Edwards. 2023. Status of
the Pacific Hake (whiting) stock in U.S. and Canadian waters in
2023. Prepared by the Joint Technical Committee of the U.S. and Canada
Pacific Hake/Whiting Agreement, National Marine Fisheries Service and
Fisheries and Oceans Canada. 208 p.
https://media.fisheries.noaa.gov/2023-02/2023-hake-assessment-post-srg_web.pdf
}
\examples{
\dontrun{
hake_recruitment
plot(hake_recruitment)
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
