% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hake_total_biomass_age_1}
\alias{hake_total_biomass_age_1}
\alias{hake_total_biomass_age_1_2024}
\title{Pacific Hake annual total biomass of age-1 fish as estimated by
the most recent (2024) stock assessment.}
\format{
A tibble also of class `pacea_biomass` with columns:
\describe{
  \item{year:}{year of the estimate of total age-1 biomass (all females and
  males), where the estimate is for the start of the year}
  \item{median:}{median estimate of age-1 biomass, in thousands of tonnes at
  the start of the year}
 }

An object of class \code{pacea_biomass} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 59 rows and 2 columns.
}
\source{
Generated from Andy running (in the hake repository)
  `pacea_save()` and then here `data-raw/groundfish/hake.R`.
}
\usage{
hake_total_biomass_age_1

hake_total_biomass_age_1_2024
}
\description{
The Pacific Hake stock is managed and assessed through an Agreement between
Canada and the United States. These estimates comes from the most
recent joint stock assessment, and, importantly, are for the coastwide stock
from California to British Columbia.
}
\details{
See `?hake_recruitment` for further details and references, and explanations
of `hake_total_biomass_age_1_2024`. Credible intervals were not easily
available but may be included in future updates.
}
\examples{
\dontrun{
hake_total_biomass_age_1
plot(hake_total_biomass_age_1)
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
