% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harbour_seals}
\alias{harbour_seals}
\title{Estimated abundance of Pacific Harbour Seals in each of seven regions}
\format{
A tibble also of class `pacea_harbour_seals` with columns:
\describe{
  \item{date:}{Date of the GAM estimate of abundance; there is a time step
  of about 250 days in the GAM, hence some years have more than one value
  and so we give a date rather than a year}
  \item{region:}{the region that the estimate corresponds to; one of SOG,
  WCVI, QCS, DP, CMC, NMC, or HG}
  \item{low:}{low end of the estimate of abundance, defined as
  the TODO CHECK}
  \item{mean:}{mean estimate of abundance, numbers of seals (in that region)}
  \item{high:}{high end of the estimate of abundance, defined TODO CHECK as
  the TODO CHECK}
 }
}
\source{
Estimates provided by Strahan Tucker, then wrangled and imported using
  `data-raw/mammals/harbour-seals/harbour-seals.R`.
}
\usage{
harbour_seals
}
\description{
The Pacific Harbour Seal (\emph{Phoca vitulina richardsi}) is the most abundant pinniped species in the
Northeast Pacific and is found throughout coastal and estuarine waters of
British Columbia. A coastwide survey of harbour seals was conducted from
2015 to 2019 and used to inform the estimated trends in abundance given here.
}
\details{
The Pacific Harbour Seal stock in Canada is
assessed as a single stock. The coastwide abundance of harbour
seals in British Columbia was assessed through aerial surveys conducted over multiple
years, and corrected for area covered, survey timing relative to peak
pupping, and proportion of seals hauled out during surveys. Estimated
abundance was calculated for seven regions: the Strait of Georgia (SOG),
West Coast Vancouver Island (WCVI), Queen Charlotte Strait (QCS), Discovery
Passage (DP), Central Mainland Coast (CMC), Northern Mainland Coast (NMC),
and Haida Gwaii (HG); see Figure 1 of DFO (2022). The estimated abundances
were calculated using Generalised Additive Models, and are included here
(yielding the trends shown in Figure 3 of DFO, 2022). See that figure to see
the amount of data available in each region; these data can be added to
`pacea` if desired -- please let us know. The final year of data in each
region is saved in the object `harbour_seals_data_final_year`.

For further details see DFO (2022), from which some of the above text was
adapted. Full details of the underlying modelling should be available in the
resulting Research Document (check
https://www.isdm-gdsi.gc.ca/csas-sccs/applications/events-evenements/result-eng.asp?DateMatch=on&StartDate=&ToDate=&mode=0&desc=PHOCA+VITULINA+RICHARDSI&mode1=0&location=&B1=Search for updates).

DFO. 2022. Stock Assessment of Pacific Harbour Seals (\emph{Phoca vitulina richardsi}) in Canada in
2019. DFO Canadian Science Advisory Secretariat Science Advisory
Report. 2022/034.
https://www.dfo-mpo.gc.ca/csas-sccs/Publications/SAR-AS/2022/2022_034-eng.html
}
\examples{
\dontrun{
harbour_seals
plot(harbour_seals)   # TODO reproduce trends in Figure 3
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
