% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harbour_seals_data_final_year}
\alias{harbour_seals_data_final_year}
\title{Final year of Pacific Harbour Seals survey in each of seven regions}
\format{
A tibble with columns:
\describe{
  \item{region:}{the region that the final year corresponds to; one of SOG,
  WCVI, QCS, DP, CMC, NMC, or HG}
  \item{final_year:}{final year of the data for that region}
 }
}
\usage{
harbour_seals_data_final_year
}
\description{
Used for colouring the plots of the estimates contained in `harbour_seals`
(since some are projections beyond the data). Also see `?harbour_seals`.
}
\examples{
\dontrun{
harbour_seals_data_final_year
}
}
\keyword{datasets}
