% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{herring_spawning_biomass}
\alias{herring_spawning_biomass}
\alias{herring_spawning_biomass_2023}
\title{Pacific Herring annual spawning stock biomass (males and females) as
estimated by the most recent stock assessment for each of the five major
stock assessment regions.}
\format{
A tibble also of class \code{pacea_herring_biomass} with columns:
\describe{
\item{year:}{year of the estimate of spawning stock biomass (males and
females)}
\item{low:}{low end (5th percentile) of the 90\\% credible interval for
biomass, units as for \code{median}}
\item{median:}{median estimate of spawning biomass, in thousands of tonnes
of spawning biomass}
\item{high:}{high end (95th percentile) of the 90\\% credible interval for
biomass, units as for \code{median}}
}

An object of class \code{pacea_biomass_herring} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 365 rows and 5 columns.
}
\source{
Generated from Andy running \code{data-raw/herring/herring.R}, which uses
model results provided by Matthew Grinnell.
}
\usage{
herring_spawning_biomass

herring_spawning_biomass_2023
}
\description{
Spawning biomass is calculated for each major stock assessment region
separately, and is calculated for males and females combined. Note that in
the assessment (DFO, 2024) spawning biomass is calculated for 2024, but that
is a projection assuming zero catch, and so is excluded here.
}
\details{
See \code{?herring_recruitment} for further details and references, and explanations
of \code{herring_spawning_biomass_2023}.
}
\examples{
\dontrun{
herring_spawning_biomass
plot(herring_spawning_biomass)
plot(herring_spawning_biomass, region = "HG")
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
