% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotssea-all-variables.R
\name{hotssea_all_variables}
\alias{hotssea_all_variables}
\title{Download all 40 HOTSSea model results from Zenodo to a local drive.}
\usage{
hotssea_all_variables(
  variables = c("temperature", "salinity"),
  run_parallel = TRUE
)
}
\arguments{
\item{variables}{character Either `c("temperature", "salinity")`,
`"temperature"`, or `"salinity"`, describing what variables to
download. Given the potentially long download times, some people may just
want temperature or salinity only.}

\item{run_parallel}{logical Run the downloads in parallel using
`parallel::foreach()`.}
}
\value{
downloaded files to `paste0(pacea_cache(), "/hotssea")`
  directory.
}
\description{
This downloads (in parallel by default) all 40 HOTSSea model results from
https://zenodo.org/records/14019142 to your local folder given by
`paste0(pacea_cache(), "/hotssea")`. Each file is around 3-9Mb, and so downloading them all may
take a while. You can check your cache folder for progress to see
them appearing there. Downloading all 40 files might take around 13 hours,
though with the parallel approach this might be only 2 hours. It depends on
network speed and number of cores.
}
\details{
Either leave running overnight, or set `variables` to "temperature" or
"salinity" if you only want one of those types of model output. Or you can just download the individual files you definitely need, see
[?hotssea_bottom_temperature_max]. Remember the downloading is a one-time
exercise (hence it's desirable to get all files at once) and then
you will have the files locally for all eternity (ish).

If this function fails, you can try rerunning it; it will not re-download any
that were successful and this worked for someone testing it.
Else you could then also try `hotssea_all_variables(run_parallel = FALSE)`, or typing `options(timeout = 1200)`
(or higher) especially if running from the Pacific Biological Station. This
function is particularly hard to test independently under different scenarios. An alternative if you are at
the Pacific Biological Station is to get Andy to copy the files to an
external drive.

Or you can just download them manually (download all) from
the Zenodo site given above) and put them in `paste0(pacea_cache(), "/hotssea")`.

Type `hotssea_data` to view the full list of available variables. See help
files for specific variables for more details on HOTSSea model results, e.g.
[?hotssea_bottom_temperature_max].

Also see the hotssea vignette.
}
\examples{
\dontrun{
hotssea_all_variables()
plot(hotssea_avg150toBot_temperature_min())
}
}
\author{
Andrew Edwards and Travis Tai
}
