% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hotssea_data}
\alias{hotssea_data}
\title{HOTSSea data layers}
\format{
table
}
\source{
Oldford et al., 2024.
Generated from running \code{data-raw/data-key/data-list.R}
}
\usage{
hotssea_data
}
\description{
List of available hotssea data layers for download.
}
\examples{
\dontrun{
hotssea_data
}

}
\keyword{datasets}
