% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isobath_200m}
\alias{isobath_200m}
\title{200m isobath for separating inshore and offshore Pacific using bathymetry}
\format{
sf multilinestring object
}
\source{
Generated from running \code{data-raw/roms/isobath_sfline.R}.
}
\usage{
isobath_200m
}
\description{
These coordinates roughly follow the continental shelf (200m depth) to provide the inshore and offshore boundary. The line is coarse and therefore ignores canyons.
}
\examples{
\dontrun{
isobath_200m
plot(isobath_200m)
plot(bccm_eez_poly, add = TRUE)
}

}
\keyword{datasets}
