% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_npi_annual}
\alias{last_npi_annual}
\title{Year-month stamp of last downloaded `npi_annual` file; not needed by users.}
\format{
character string
}
\source{
Generated and used in `data-raw/coastwide-indices/coastwide-indices.R`.
}
\usage{
last_npi_annual
}
\description{
Used in `data-raw/coastwide-indices/coastwide-indices.R` because each updated index is given a
new website address.
}
\author{
Andrew Edwards
}
\keyword{datasets}
