% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_npi_monthly}
\alias{last_npi_monthly}
\title{Year-month stamp of last downloaded \code{npi_monthly} file; not needed by users.}
\format{
character string
}
\source{
Generated and used in \code{data-raw/coastwide-indices/coastwide-indices.R}.
}
\usage{
last_npi_monthly
}
\description{
Used in \code{data-raw/coastwide-indices/coastwide-indices.R} because each updated index is given a
new website address.
}
\author{
Andrew Edwards
}
\keyword{datasets}
