% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-grid.R
\name{make_grid}
\alias{make_grid}
\title{Make the master grid}
\usage{
make_grid(coastline, cell_size_x = 2, cell_size_y = 2, rotang = 318.5)
}
\arguments{
\item{coastline}{`SpatialPolygons` object (from `sp` package) to represent
the coastline. Probably only works for `Coastline` for now, which Joe got
from somewhere.}

\item{cell_size_x}{size of each cell in the (pseudo, due to rotation) x
direction, km}

\item{cell_size_y}{size of each cell in the (pseudo, due to rotation) y
direction, km}

\item{rotang}{angle to rotate the grid}
}
\value{
a grid to use as a master grid, of class `sf` and `data.frame`
}
\description{
Internal function for developers, to create the master grid. Andy
functionalised from Joe's original code. Function will need
changing/adapting/deprecating if we use a new coastline object. And if we
decide on a non-regular grid - coarser away from the coast.
}
\examples{
\dontrun{
# See data-raw/grids/test-grids.R
}
}
\author{
Andrew Edwards
}
