% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npgo}
\alias{npgo}
\title{NPGO -- North Pacific Gyre Oscillation}
\format{
A tibble also of class \code{pacea_index} with columns:
\describe{
\item{year:}{year of value}
\item{month:}{month (1 to 12) of value}
\item{anomaly:}{calculated as the second dominant mode of sea-surface height
anomaly over the Northeast Pacific (180 deg W to 100 deg W, 25 deg N to 62
deg N), and then maybe converted to standard deviation units as in Figure
1 of Di Lorenzo et al. (2008), though the numbers do not exactly match
that figure (and come from Di Lorenzo's website, which
notes that values after 2004 were updated. This might be why
numbers do not seem to quite agree.}
}
}
\source{
Generated from running \code{data-raw/coastwide-indices/coastwide-indices.R}.
}
\usage{
npgo
}
\description{
The North Pacific Gyre Oscillation (NPGO) is a climate pattern that
is significantly correlated with fluctuations of salinity, nutrients, and chlorophyll-a
in long-term observations in the California Current (CalCOFI) and Gulf of
Alaska (Line P).
}
\details{
Adapted from http://www.o3d.org/npgo/
The NPGO emerges
as the second dominant mode of sea-surface height variability in the
Northeast Pacific. The term NPGO is used because fluctuations reflect
changes in the intensity of the central and eastern branches of the North
Pacific gyre circulations as evident from the NPGO sea-surface height
anomalies. Fluctuations in the NPGO are driven by regional and basin-scale
variations in wind-driven upwelling and horizontal advection -- the
fundamental processes controlling salinity and nutrient
concentrations. Nutrient fluctuations drive concomitant changes in
phytoplankton concentrations, and may force similar variability in higher
trophic levels. The NPGO thus provides a strong indicator of fluctuations in
the mechanisms driving planktonic ecosystem dynamics.

Key reference is Di Lorenzo et al. (2008): North Pacific Gyre Oscillation
links ocean climate and ecosystem change. Geophysical Research Letters, 35:
L08607. doi:10.1029/2007GL032838

Associated code adapted from code generously shared by Chris Rooper.
}
\examples{
\dontrun{
npgo
plot(npgo)
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
