% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npi_annual}
\alias{npi_annual}
\title{NPI -- North Pacific Index annual values and anomalies}
\format{
A tibble also of class `pacea_index` with columns:
\describe{
  \item{year:}{year of value}
  \item{val:}{absolute average winter value, hPa. The value for year $N$
  refers to an average of the values in November  $N-1$, December $N-1$, and
  January, February, and March for year $N$. For example, the 1999 annual
  value contains the average of November and December 1998 and January,
  February, and March 1999. Hence there is an `NA` for 1899 as no monthly
  data for 1898.}
  \item{anom:}{anomalies of the annual winter values compared to the 1925-1989 mean
  of 1008.9 hPa, again with an `NA` for 1899 TODO check.}
}
}
\source{
Generated from running `data-raw/coastwide-indices/coastwide-indices.R`.
}
\usage{
npi_annual
}
\description{
The North Pacific index measures interannual to decadal
variations in atmospheric circulation across the North Pacific.
}
\details{
See `?npi_monthly` for details.

`npi_annual` contains annual average winter values and anomalies.
}
\examples{
\dontrun{
npi_annual
plot(npi_annual)
}

}
\author{
Andrew Edwards
}
\keyword{datasets}
