% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npi_monthly}
\alias{npi_monthly}
\title{NPI -- North Pacific Index monthly values}
\format{
A tibble also of class `pacea_t` (pacea temporal) with columns:
\describe{
  \item{year:}{year of value}
  \item{month:}{month (1 to 12) of value}
  \item{val:}{absolute monthly value (average?), hPa, with an `NA` for
  December 1944}
 }
}
\source{
Generated from running `data-raw/coastwide-indices/coastwide-indices.R`.
}
\usage{
npi_monthly
}
\description{
The North Pacific index measures interannual to decadal
variations in atmospheric circulation across the North Pacific.
}
\details{
The North Pacific Index (NPI) is the area-weighted sea level
pressure over the region 30 deg N to 65 deg N, 160 deg E to 140 deg W. The
dominant atmosphere-ocean relation in the North Pacific is one where
atmospheric changes lead changes in sea surface temperatures (SST) by one to two
months. However, strong ties exist with events in the tropical Pacific, with
changes in tropical Pacific SSTs leading SSTs in the north Pacific by three
months. Taken from
https://climatedataguide.ucar.edu/climate-data/north-pacific-np-index-trenberth-and-hurrell-monthly-and-winter
Key reference is Trenberth and Hurrell (1994): Decadal atmosphere-ocean
variations in the Pacific, Climate Dynamics 9:303-319.

`npi_monthly` contains monthly absolute values. Annual average winter values
and anomalies are in `npi_annual`.
Ross and Robert (2022; see `?oni` for reference): The NPI is a useful
indicator of the intensity and real extent of the Aleutian Low Pressure
system. The NPI was generally positive from 1950 to 1976, and generally
negative (red) from 1977 to 2008; a change than can be attributed to the
strengthening of the Aleutian Low Pressure system after 1977. From 2008 to
present [2021], the NPI was mostly positive, due to weaker Aleutian
Lows. The NPI anomaly, plotted in Figure 7-2 [of Ross and Robert (2022)],
was calculated from the NPI by removing the 1950-2018 mean.

Associated code adapted from code generously shared by Chris Rooper.
}
\author{
Andrew Edwards
}
\keyword{datasets}
