% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oisst_7day}
\alias{oisst_7day}
\alias{oisst_month}
\title{Optimal interpolation SST}
\format{
\describe{
  \item{year}{Year}
  \item{week}{Calendar week}
  \item{month}{Calendar month}
  \item{sst}{Mean sea surface temperature, SST (deg C)}
  \item{sst_sd}{Standard deviation of SST}
  \item{sst_n}{Sample size of SST}
  \item{start_date}{Start date of mean SST estimate}
  \item{end_date}{End date of mean SST estimate}
}
`oisst_7day` An object of class sf with 2051650 rows and 8 columns.

`oisst_month` An object of class sf with 464350 rows and 8 columns.
}
\source{
https://www.ncei.noaa.gov/products/optimum-interpolation-sst

Generated from running `data-raw/erddap/OISST.R`.
}
\usage{
oisst_7day

oisst_month
}
\description{
Simple features objects of NOAA's Optimal Interpolation SST
\describe{
  \item{oisst_7day}{7-day mean sea surface temperature (deg C)}
  \item{oisst_7dayclim}{7-day mean climatology (1991-2020) for sea surface temperature (deg C)}
  \item{oisst_month}{Monthly mean sea surface temperature in (deg C}
  \item{oisst_monthclim}{Monthly climatology (1991-2020) for sea surface temperature (deg C)}
}
}
\details{
From NOAA:
"The NOAA 1/4° Daily Optimum Interpolation Sea Surface Temperature (OISST) is a long term Climate Data Record that incorporates observations from different platforms (satellites, ships, buoys and Argo floats) into a regular global grid. The dataset is interpolated to fill gaps on the grid and create a spatially complete map of sea surface temperature. Satellite and ship observations are referenced to buoys to compensate for platform differences and sensor biases."

The data starts from 1981-09-01 and ends in the previous month from the current month. These data will be updated monthly.
}
\examples{
\dontrun{
head(oisst_7day)
head(oisst_7dayclim)
}

}
\keyword{datasets}
