% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oni}
\alias{oni}
\title{ONI -- Oceanographic Niño Index}
\format{
A tibble also of class \code{pacea_index} with columns:
\describe{
\item{year:}{year of value}
\item{month:}{month (1 to 12) of value}
\item{val:}{absolute values of three-month averages (preceding, current, and
next month), deg C; note that recent values may change in subsequent
updates -- see details}
\item{anom:}{anomalies based on 30-year base periods that are updated every
5 years, deg C; note that recent values may change in subsequent
updates -- see details}
}
}
\source{
Generated from running \code{data-raw/coastwide-indices/coastwide-indices.R}.
}
\usage{
oni
}
\description{
The Oceanic Niño Index is a monthly index which is one measure of the El
Niño-Southern Oscillation.
}
\details{
NOTE (2nd March 2026). Now we are in 2026, there is a new 30-year baseline
period available for calculating anomalies, from 1995-2025. This is now used
for all all years from 2010 onwards; hence such values will have changed from previous versions of
pacea. See below for more details.

The Oceanic Niño Index (ONI) is a 3-month running
mean of sea surface temperature (SST) anomalies in the Niño 3.4 region
(5 deg N to 5 deg S, 120 deg W to 170 deg W) plotted on
the centre month. The SST anomalies are calculated based on 30-year base
periods that are updated every 5 years, which accounts for global warming
and some of the decadal-scale SST variability (as seen in the Pacific
Decadal Oscillation index).
The ONI is provided by the NOAA’s National Weather
Service National Centers for Environmental Prediction CPC:
http://www.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/ensoyears.shtml

Preceding paragraph adapted from: Ross, T., and Robert, M. (2022). Normal
temperatures despite strong cool
climate indices and an emerging freshening trend. Pages 23-30 of
Boldt, J.L., Joyce, E., Tucker, S., and Gauthier, S. (Eds.). 2022. State of
the physical, biological and selected fishery resources of Pacific Canadian
marine ecosystems in 2021. Canadian Technical Report of Fisheries and
Aquatic Sciences. 3482 vii+242 p.

The above website is updated automatically on the first Thursday of each
month, and states that:

Because of the high frequency filter applied to the
ERSSTv5 data, ONI values may change up to two months after
the initial "real time" value is posted. Therefore, the most recent ONI
values should be considered an estimate.
On the site, Warm (red) and cold (blue) periods are based on a threshold of
+/- 0.5 deg C for the Oceanic Niño Index (ONI), using a 3-month running mean of
ERSST.v5 SST anomalies in the Niño 3.4 region (5 deg N-5 deg S, 120 deg W
-170 deg W), based on centered 30-year base periods updated every 5 years.
For historical purposes, periods of below and above normal SSTs are colored in
blue and red (on the website) when the threshold is met for a minimum of 5
consecutive overlapping seasons. The ONI is one measure of the El
Niño-Southern Oscillation, and other indices can confirm whether features
consistent with a coupled ocean-atmosphere phenomenon accompanied these periods.

Also see https://www.ncei.noaa.gov/access/monitoring/enso/sst

2nd March 2026 update: The 'centered 30-year base period updated every 5 years' is somewhat
nuanced. The following is based on this 2013 NOAA web article:
https://www.climate.gov/news-features/understanding-climate/watching-el-ni\%C3\%B1o-and-la-ni\%C3\%B1a-noaa-adapts-global-warming:

Each five-year period in the historical record now has its own
30-year average centered on the first year in the period: the years 1950 to
1955 are compared to the 1936-1965 average, for example, while the years
1956-1960 are compared to 1941-1970.

For years for which a 30-year period is
not available (because the final years have not occurred yet), the
the most recently calculated climatology is used.

So, for example:
\enumerate{
\item In 2025, the oni values from 2010 onwards would have been based on the
climatology from 1990-2020, because the years 1996-2025 would not be
available (the full temperatures throughout 2025 would not yet be known),
and so 1990-2020 is the last complete 30-year block available that ends in a
year that is a multiple of 5 years.
\item In March 2026, when updating pacea, there is a new 30-year block available,
1995-2025. This is a 30-year block centred on 2010, so this is now used
as the baseline for all years from 2010-2014.
\item Now that 1995-2025 is the newest 30-year block available, it is also used for
all years from 2015 onwards.
}

Thus all oni values from 2010 onwards have changed when updating pacea in
March 2026, compared to versions processed in 2025 and earlier. Pre-2010
values have not changed because the relevant 30-year averages have not changed.
Remember to use \code{pacea_installed()} to see when you last installed pacea.

Ocean temperatures have been increasing. The reason for calculating
anomalies with respect to near-in-time 30-year averages, rather than the
same 30-year average for the full time series, is described
by Mike Halpert at NOAA's Climate Prediction Center:

If using one 30-year average for the full time series "we could picture a
day not too far in the future when a weak El Niño episode from the past
would wind up looking like a weak La Niña episode when compared to the more
recent, warmer ocean temperatures."

Associated code adapted from code generously shared by Chris Rooper.
}
\examples{
\dontrun{
oni
plot(oni)
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
