% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacea-import-from.R
\docType{package}
\name{pacea}
\alias{pacea-package}
\alias{pacea}
\title{pacea: an R package of Pacific ecosystem information to help facilitate an
ecosystem approach to fisheries management}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains a variety of data sets and model output. We wrangle the data sets behind the scenes to get them into usable formats in R, and provide helpful plotting functions. All data sets include documentation regarding the original sources and authors, who should be cited and consulted as appropriate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pbs-assess/pacea}
  \item Report bugs at \url{https://github.com/pbs-assess/pacea/issues}
}

}
\author{
\strong{Maintainer}: A. M. Edwards \email{andrew.edwards@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0003-2749-8198}{ORCID})

Authors:
\itemize{
  \item T. C. Tai \email{travis.tai@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0001-7013-1901}{ORCID})
  \item J. Watson (\href{https://orcid.org/0000-0002-4773-1631}{ORCID})
  \item M. A. Peña \email{angelica.pena@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0001-8002-3402}{ORCID})
  \item A. Hilborn \email{andrea.hilborn@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-5967-5788}{ORCID})
  \item C. G. Hannah \email{charles.hannah@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-0238-5683}{ORCID})
  \item C. N. Rooper \email{chris.rooper@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0003-2315-1269}{ORCID})
  \item K. L. Flynn \email{kelsey.flynn@dfo-mpo.gc.ca}
  \item G. L. Oldford \email{greig.oldford@dfo-mpo.gc.ca}
}

Other contributors:
\itemize{
  \item Jessica Nephin  [contributor]
  \item Strahan Tucker [contributor]
  \item Matt Grinnell  [contributor]
  \item Lindsay Davidson  [contributor]
  \item Sean Anderson [contributor]
  \item Philina English [contributor]
  \item Chris Grandin (\href{https://orcid.org/0000-0002-6746-3486}{ORCID}) [contributor]
  \item Jennifer Boldt [contributor]
  \item Kelly Young [contributor]
}

}
\keyword{internal}
