% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacea-cache.R
\name{pacea_cache}
\alias{pacea_cache}
\title{Get directory of where to cache large data sets}
\usage{
pacea_cache()
}
\value{
the cache directory on the user's machine, based on
\code{rappdirs::user_cache_dir("pacea")}. On Windows in 2023 this seems to be
\verb{C:\\\\Users\\\\<username>\\\\AppData\\\\Local\\\\pacea\\\\Cache}. If you want to
put use something else then see \code{?rappdirs::user_cache_dir} and change
your \code{R_USER_CACHE_DIR}. To do that see
\verb{https://github.com/r-lib/rappdirs/issues/34}.
}
\description{
Some data sets are too large to save as data objects within pacea and need
to be cached locally; this gives the default location.
}
\details{
Some data sets need to be cached locally and will not be needed by all users
(and so do not really need to be stored as R data objects in the
project). Also, some large ones may be regularly updated. When such data are
needed, we want to cache them locally rather than download them from an
external website each time pacea is used in a new R session, and make them
easy to update.
}
\examples{
\dontrun{
pacea_cache()
}
}
\author{
Andrew Edwards
}
