% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_cache.R
\name{pacea_clear_cache}
\alias{pacea_clear_cache}
\title{ROMS clear cache}
\usage{
pacea_clear_cache()
}
\description{
Delete all ROMS data files in cache folder
}
\examples{
\dontrun{
get_pacea_data("roms_surface_temperature")
}

}
