% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacea_data}
\alias{pacea_data}
\title{ROMS data layers}
\format{
table
}
\source{
Pena et al., 2019.
}
\usage{
pacea_data
}
\description{
List of available ROMS data layers for download.
}
\examples{
\dontrun{
pacea_data
}

}
\keyword{datasets}
