% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacea-import-from.R
\docType{package}
\name{pacea_import_from}
\alias{pacea_import_from}
\title{pacea package importFrom requirements}
\description{
Adapting from https://github.com/pbs-assess/gfiphc/blob/master/R/gfiphc.R
(see that if get issues with column names within dplyr functions for what to
 add). And from
https://github.com/andrew-edwards/sizeSpectra/blob/master/R/sizeSpectra.R for
some tibble related stuff.
And from the 'Consider adding' output from check(vignettes = FALSE), adding
here to then go into NAMESPACE
}
