% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacea_indices}
\alias{pacea_indices}
\title{Summary of climatic and oceanographic indices in pacea}
\format{
A tibble with columns:
\describe{
  \item{Object:}{name in pacea (use ?<object> for help file}
  \item{Description:}{Long name}
  \item{Resolution:}{Monthly or annual resolution}
  \item{Start year:}{Start year of time series}
  \item{End year:}{End year of time series}
}
}
\source{
Generated from running `data-raw/coastwide-indices/coastwide-indices.R`.
}
\usage{
pacea_indices
}
\description{
The Aleutian Low Pressure Index (ALPI) has been used to describe
decadal-scale changes in North Pacific climate-ocean conditions and is linked
to patterns in marine productivity.
}
\examples{
\dontrun{
pacea_indices
knitr::kable(pacea_indices)   # in vignette and README
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
