% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacea-wrangle.R
\name{pacea_long}
\alias{pacea_long}
\alias{pacea_wide}
\title{pacea spatiotemporal data convert to long/wide format}
\usage{
pacea_long(data, names_to = "date", values_to = "value")

pacea_wide(data, names_from = c("year", "month"), values_from = "value")
}
\arguments{
\item{data}{pacea.st data object in long format to pivot.}

\item{names_to}{A character value for the column name representing the date. Defaults to separate 'year' and 'month' for BCCM data.}

\item{values_to}{A character value for the column name representing variable data (e.g. temperature)}

\item{names_from}{A character vector representing the date from which to name columns from. If vector is greater than 1, column values will be pasted together, separated by '_'. Defaults to paste BCCM data in 'year_month' format.}

\item{values_from}{A character value representing the column names from which to get variable data values.}
}
\value{
pacea spatiotemporal data in long/wide format
}
\description{
Simplified function for users, taking the \code{tidyr::pivot_longer} and \code{tidyr::pivot_wider} functions to transform pacea.st data into long or wide format for simpler data manipulation, summarising, and plotting. Writted to only work with 'pacea.st' (spatiotemporal) data.
}
\examples{
\dontrun{
pdata <- bccm_surface_temperature()[,1:2]
pdata_long <- pacea_long(pdata)
head(pdata_long)
}
\dontrun{
pdata <- bccm_surface_temperature()[,1:2]
pdata_long <- pacea_long(pdata)
pdata_wide <- pacea_wide(pdata)
head(pdata_wide)
}
}
