% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pacific_Herring_Sections.R
\docType{data}
\name{Pacific_Herring_Sections}
\alias{Pacific_Herring_Sections}
\title{Contains the Definitions of 108 Pacific Herring Statistical Sections}
\format{
An sf Polygons object:
}
\usage{
Pacific_Herring_Sections
}
\description{
A sf object
}
\keyword{datasets}
