% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfma_area_126_temp}
\alias{pfma_area_126_temp}
\title{Temporary PFMA Area 126 sf object for testing area options in code}
\format{
A simple features polygon dataframe.
}
\source{
Generated from running `data-raw/areas/pfma-area-126-temp.R"
}
\usage{
pfma_area_126_temp
}
\description{
A simple features object of PFMA Area 126 to test code.
}
\examples{
\dontrun{
pfma_area_126_temp
plot(pfma_area_126_temp)    # just draws the polygon of the area
# Plot the latest OISST data plus draw the polygon of the area
p <- plot(oisst_month)

p + geom_sf(data = pfma_area_126_temp,
            fill = NA,
            colour = "black",
            linewidth = 1)
}

}
\keyword{datasets}
