% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-biomass-herring.R
\name{plot.pacea_biomass_herring}
\alias{plot.pacea_biomass_herring}
\title{Plot a herring biomass time series object for one or all herring regions}
\usage{
\method{plot}{pacea_biomass_herring}(
  obj,
  region = NULL,
  x_lab = "Year",
  y_lab = attr(obj, "axis_name"),
  mar_all_regions = c(3, 3, 2, 0),
  oma_all_regions = c(2, 1, 1, 1),
  y_tick_by = 5,
  title = "full",
  ...
)
}
\arguments{
\item{obj}{a \code{pacea_biomass_herring} object, which is a tibble of
annual spawning biomass or total biomass estimates (TODO not yet
implemented) for each of the five herring Stock Assessment Regions.}

\item{region}{if \code{NULL} (the default) then plots a panel plot of biomass
for all five regions, else plots the specified region, which must be one
of HG, PRD, CC, SOG, or WCVI.}

\item{x_lab}{label for x-axis (default is "Year", only shown for bottom
panel when plotting all five regions)}

\item{y_lab}{label for y-axis (default is specified by \code{attr(obj, "axis_name")}.}

\item{mar_all_regions}{\code{mar} value for \code{par()}, may need tweaking for
five-panel plot; for single-panel just specific \code{mar}; see \code{?par}}

\item{oma_all_regions}{similar to \code{mar_all_regions}}

\item{y_tick_by}{increment for y-axis ticks; gets
overwritten in \code{add_tickmarks()} if this yields more than
\code{y_tick_max_number} tickmarks. If using \code{\link[=plot.pacea_biomass]{plot.pacea_biomass()}} the default
of 1 gets automatically changed to 0.25 for \code{plot(hake_biomass)}. If
\code{y_axis_reverse} is TRUE then the negative of \code{y_tick_by} (which should
always be positive) is passed onto \code{add_tickmarks()}.}

\item{title}{if \code{NULL} then no figure title (using \code{main()}), if \code{full} (the
default) then spell out the Stock Assessment Region, and if \code{short} then
just use the acronym.}

\item{...}{further options passed onto \code{plot.pacea_biomass()} that can
also pass onto \code{plot.default()}}
}
\value{
plot of the herring biomass time series as joined up medians with
shaded regions showing uncertainty, for either all regions or the
specified region.
}
\description{
Adapted slightly from \code{plot.pacea_recruitment_herring()}. Plots in the
biomass style developed in \code{plot.pacea_biomass()}.
}
\examples{
\dontrun{ TODO

}
}
\author{
Andrew Edwards
}
