% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-buoy-sst.R
\name{plot.pacea_buoy}
\alias{plot.pacea_buoy}
\title{Plot buoy SST data for one or multiple stations}
\usage{
\method{plot}{pacea_buoy}(
  obj,
  stn_id = "C46146",
  years = NULL,
  year_highlight = lubridate::year(lubridate::today())
)
}
\arguments{
\item{obj}{buoy SST data, of class `pacea_buoy_sst`}

\item{stn_id}{single station to plot, or multiple station ids (as given by
`stn_id` column in `buoy_metadata`). Or "all" to plot all stations. TODO}
}
\description{
Plot buoy SST data for either one station or multiple stations, with various options.
}
\examples{
\dontrun{
}
}
\author{
Andrew Edwards
}
