% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-indices.R,
%   R/plot-pacea-buoy-anomalies-list.R, R/plot-pacea-buoy-anomalies-single.R,
%   R/plot-pacea-indices.R
\name{plot_indices}
\alias{plot_indices}
\alias{plot.pacea_buoy_anomalies_list}
\alias{plot_pacea_buoy_anomalies_single}
\alias{plot_pacea_indices}
\title{Plot multiple indices, that have been created by the user, together in a similar heatmap style to the buoy SST plot.}
\usage{
plot_indices(
  ...,
  years = NULL,
  main = NULL,
  xlab = "Year",
  ylab = "Index",
  return_results = FALSE,
  scale_limits = NULL,
  angle_year_labels = TRUE,
  gap_after = NULL,
  gap_size = 0.2
)

\method{plot}{pacea_buoy_anomalies_list}(
  pacea_buoy_anomalies_list,
  stn_id_to_plot = NULL,
  months = NULL,
  main = NULL,
  xlab = "Year",
  ylab = "Buoy",
  use_stn_id_name = TRUE,
  sst_plot = "anomalies",
  count_breaks = NULL,
  return_results = FALSE,
  require_requested_months = NULL,
  scale_limits = NULL
)

plot_pacea_buoy_anomalies_single(
  pacea_buoy_anomalies_list,
  stn_id_to_plot,
  months,
  main,
  xlab,
  ylab,
  use_stn_id_name,
  sst_plot = "anomalies",
  count_breaks = NULL,
  return_results = FALSE,
  scale_limits = NULL
)

plot_pacea_indices(
  ...,
  months = NULL,
  years = 1970:(lubridate::year(lubridate::today())),
  main = NULL,
  xlab = "Year",
  ylab = "Index",
  return_results = FALSE,
  require_requested_months = NULL,
  scale_limits = NULL
)
}
\arguments{
\item{...}{the indices you want to plot, e.g. oni, pdo, etc., any of
those listed in \code{pacea_indices}.}

\item{years}{numeric vector of the years to show (note that PDO goes back to
1854 to we selected something more recent for the default start value);
default end is the current year}

\item{main}{title for the plot, if \code{NULL} then created automatically,
including detailing the months selected. May need to manually specify \code{main}
if many non-consecutive months are chosen (which seems unlikely).}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{return_results}{logical, if \code{FALSE} (default) returns the plot object only.
If \code{TRUE}, prints the plot and returns a list with both \code{plot} and \code{results}.}

\item{scale_limits}{numeric vector of length 2 specifying the limits for the colour scale.
If \code{NULL} (default), limits are calculated automatically based on the data. Only used
for non-count plots. If any plotted values are outside of \code{scale_limits}
then they are coloured grey.}

\item{angle_year_labels}{logical, whether to put the year labels at 45
degrees, which is needed to avoid overlapping when there are many
years. TODO maybe need the next rdname:}

\item{gap_after}{character vector (or NULL, the default) of index names
to add a gap after in the plot. For example,
\code{gap_after = "herring_recruitment_sog_index"} (which is an index you have
created and named using \code{create_index()}) will add a gap after that index.
Multiple indices can be specified: \code{gap_after = c("index1", "index2")}.}

\item{gap_size}{numeric value controlling the size of the gaps in cm
(default 0.2, roughly 1/5 of the box height).}

\item{pacea_buoy_anomalies_list}{object of class \code{pacea_buoy_anomalies_list}
obtained from running \code{calculate_anomalies()} on buoy data.}

\item{stn_id_to_plot}{single \code{stn_id} to plot}

\item{months}{numeric vector of months (1-12) to include. If not specified
then defaults to 4
(April), except when only one \code{stn_id_to_plot} when all months are plotted (unless specified). Can have more than one month, e.g. 6:9. For a 'winter' average,
say Nov-Mar, specify the months as \code{c(11, 12, 1, 2, 3)}. The winter average
anomaly will be calculated and named for the year in which January
falls. \code{months} must be consecutive (except for the overwintering \verb{12, 1} as
in the above example).}

\item{use_stn_id_name}{logical, whether to use the name of the buoy
(e.g. \verb{Middle NOMAD}) or the \code{stn_id} (e.g. C46004). See \code{buoy_metadata} for
the names.}

\item{sst_plot}{character, one of:
\itemize{
\item \code{"anomalies"} (default) - plots SST anomalies
\item \code{"mean"} - plots mean SST values for each month
\item \code{"count"} - plots count of daily SST values used to calculate monthly mean
}}

\item{count_breaks}{numeric vector of break points for the count plot colour scale.
Only used when \code{sst_plot = "count"}. Default of \code{NULL} actually defaults to
\code{c(0, 10, 15, 20, 31) * length(months)} if \code{months} is specified, else it is
\code{c(0, 10, 15, 20, 31)}.}

\item{require_requested_months}{numeric, number of months that must be available
out of the requested \code{months} to compute an average anomaly. If \code{NULL} (default),
defaults to \code{length(months)}. Only used when \code{length(months) > 1} and
\code{length(stn_id_to_plot) > 1}.}
}
\value{
a ggplot object (when \code{return_results = FALSE}) or a list with \code{plot} and \code{results}
(when \code{return_results = TRUE})

a ggplot object (when \code{return_results = FALSE}) or a list with \code{plot} and \code{results}
(when \code{return_results = TRUE})

a ggplot object

a ggplot object (when \code{return_results = FALSE}) or a list with \code{plot} and \code{results}
(when \code{return_results = TRUE})
}
\description{
Takes any number standardised annual indices created using \code{\link[=create_index]{create_index()}} and
plots a heatmap style plot.

TODO Need to write this.

Takes any number of the climatological and oceanographic indices, such as those in
\code{pacea_indices}. For plotting standard time series of one index just use
\code{\link[=plot]{plot()}} which then calls \code{\link[=plot.pacea_index]{plot.pacea_index()}}.
}
\details{
To display square tiles for each data point, manually adjust your figure
dimensions when saving.

Not currently set up to use \code{npi_monthly} and \code{npi_annual} as
they are absolute values and so need a climatology defined (contact Andrew
Edwards if you want that implemented; they also have the only \code{NAs}, namely
Dec 1944 and 1899 respectively). Note that \code{bi} and \code{alpi} are annual
values (others are monthly), and so only get shown if \code{months} is
\code{1:12}. Also, since \code{alpi} has absolute units (of millions of square km) it
is normalised here (subtract the mean and divide by the standard deviation,
calculate over the years selected). Though other indices are not normalised
in this way -- this is likely something to be discussed with a wider audience.
}
\examples{
\dontrun{
# See the vignette for explanations and use of the options.
hake_recruitment_index <- create_index(hake_recruitment)
herring_wcvi_recruitment_index <- create_index(herring_recruitment)   # default is WCVI
plot_indices(hake_recruitment_index,
             herring_wcvi_recruitment_index)
# hake_rec_over_2010_index <- create_index(hake_recruitment_over_2010,
# index_label = "Rec over 2010")   # Just to get it working, values should be
# close TODO think about why 2021 values don't match, others seem to


}
\dontrun{
# See the vignette for explanations and use of the options.
all_buoys_anomalies <- calculate_anomalies(buoy_sst,
                                           climatology_time = "month")
plot(all_buoys_anomalies)
}
\dontrun{
# See the vignette
}
\dontrun{
# See the vignette for explanations and use of the options.
plot_pacea_indices(oni, pdo, mei)
xx <- plot_pacea_indices(oni, pdo, mei, alpi, bi, months = 1:12, return_results = TRUE)
xx
}
}
\author{
Andrew Edwards and Andrea Hilborn

Andrew Edwards
}
