% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-buoy-anomaly-list.R
\name{plot.pacea_buoy_anomaly_list}
\alias{plot.pacea_buoy_anomaly_list}
\title{Plot anomalies of the buoy sea-surface temperature data}
\usage{
\method{plot}{pacea_buoy_anomaly_list}(
  pacea_buoy_anomaly_list,
  stn_id_to_plot = NULL,
  months = NULL,
  main = NULL,
  xlab = "Year",
  ylab = "Buoy",
  use_stn_id_name = TRUE
)
}
\arguments{
\item{pacea_buoy_anomaly_list}{object of class \code{pacea_buoy_anomaly_list}
obtained from running \code{caclulate_anomaly()} on buoy data.}

\item{stn_id_to_plot}{character vector of station IDs (\code{stn_id} values)from the \code{buoy_sst} data
object. If \code{NULL} (default), anomalies for all buoys (in
\code{pacea_buoy_anomaly_list_object}) are plotted. If specified, only the
buoys in this vector are plotted. If \code{length(stn_id_to_plot == 1)} then anomalies
for each month are shown, with January at the top and December at the
bottom. TODO could make it accept the names instead.}

\item{months}{numeric vector of months to include (1-12). If not specified
then defaults to 4
(April), except when only one \code{stn_id_to_plot} when all months are plotted (unless specified). Can have more than one month, e.g. 6:9. For a 'winter' average,
say Nov-Mar, specify the months as \code{c(11, 12, 1, 2, 3)}. The winter average
anomaly will be calculated and named for the year in which January falls. TODO}

\item{main}{title for the plot, if \code{NULL} then created automatically,
including detailing the months selected. May need to manually specify \code{main}
if many non-consecutive months are chosen (which seems unlikely).}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{use_stn_id_name}{logical, whether to use the name of the buoy
(e.g. \verb{Middle NOMAD}) or the \code{stn_id} (e.g. C46004). See \code{buoy_metadata} for
the names.}
}
\value{
a ggplot object
}
\description{
TODO Need to clearly explain methods and run them by someone. And put checks
in for there being enough data in each month. e.g. months = 12 has empty for
C46132 but not for months 11:12, in year 2015. Get code working then figure
out the many caveats.
}
\examples{
\dontrun{
all_buoys_anomalies <- calculate_anomaly(buoy_sst,
                             climatology_time = "month")
all_buoys_plot <- plot.pacea_buoy_anomaly_list(all_buoys_anomalies)
all_buoys_plot

all_buoys_plot <- plot.pacea_buoy_anomaly_list(all_buoys_anomalies, months =
1:4)
# TODO figure out that
# all_buoys_plot <- plot(all_buoys_anomalies, months = 1:12)
# and
# all_buoys_plot_2 <- plot(all_buoys_anomalies, months = 12:1)
# are correctly different. Latter should really give an error.
# TODO currently gives the title correctly, but need to change the function
}
}
\author{
Andrew Edwards
}
