% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-harbour-seals.R
\name{plot.pacea_harbour_seals}
\alias{plot.pacea_harbour_seals}
\title{Plot the Pacific Harbour Seals abundance estimates}
\usage{
\method{plot}{pacea_harbour_seals}(
  obj,
  region = NULL,
  value = "mean",
  include_coastwide = TRUE,
  xlab = "Year",
  ylab = attr(obj, "axis_name"),
  y_tick_by = 1000,
  x_tick_extra_years = 20,
  start_decade_ticks = lubridate::ymd("1800-01-01", truncated = 2),
  y_max = NULL,
  uncertainty_shade_col = rgb(0, 0, 1, 0.1),
  uncertainty_line_col = "blue",
  uncertainty_line_lty = 3,
  median_type = "l",
  median_pch = 16,
  median_line_col = "black",
  median_line_lty = 1,
  median_line_lwd = 2,
  ...
)
}
\arguments{
\item{obj}{`pacea_harbour_seals`, as this is of class `pacea_harbour_seals`
and the plotting is tailored for this data set.}

\item{region}{which region to plot for a single plot; must be one of SOG,
WCVI, QCS, DP, CMC, NMC, HG, or coastwide. If `NULL` (the default) then do a panel
plot showing all seven regions.}

\item{value}{the column to plot as a solid line, in this case always the
default `mean`}

\item{include_coastwide}{logical, whether to include the coastwide population in the
eight panel plot (keep `region = NULL`).}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label, the default is an attribute of the `pacea_index`
object.}

\item{y_tick_by}{increment for y-axis ticks}

\item{x_tick_extra_years}{number of extra years to expand around the range
of data for which to add annual tick marks (does not expand the axis); in
hindsight could have simplified this in `add_tickmarks()`, but just made
the default big here.}

\item{start_decade_ticks}{where to start tickmarks for decades (defaults to
1800 as hard to automate)}

\item{y_max}{maximum y value for certain types of plot (use this if you get
an error when specifying `ylim`)}

\item{uncertainty_shade_col}{colour of shading for uncertainty}

\item{uncertainty_line_col}{colour of line for uncertainty}

\item{uncertainty_line_lty}{lty of line for uncertainty}

\item{median_type}{`type` of plot ("o", "p", etc.) for means (note: we have
used median elsewhere so kept with that name, but the seal estimates are
means not medians)}

\item{median_pch}{pch for median}

\item{median_line_col}{col for median}

\item{median_line_lty}{lty for median}

\item{median_line_lwd}{lwd for median}

\item{...}{optional arguments passed onto `plot()`. Note that the x-axis is
constructed using a lubridate `date` object, so `xlim` needs to be a
`date` object (see example).}
}
\value{
plot of the time series as median with bars showing uncertainty (if
  `low` and `high` are columns of `obj) to the current device; returns nothing.
}
\description{
Default is to plot all seven regions plus the coastwide estimate, to mostly replicate
Figure 3 of DFO (2022; see `?harbour_seals`) in a `ggplot`
style. If a single region is
indicated then a time series plot is shown, in the same style as
plots of object class `pacea_biomass` (such as `hake_biomass`).
Note that the values are means and confidence intervals (not medians and
confidence intervals like other objects); see `?harbour_seals`.
}
\examples{
\dontrun{
plot(harbour_seals)
plot(harbour_seals, region = "SOG")
plot(harbour_seals, include_coastwide = FALSE)
}
}
\author{
Andrew Edwards
}
