% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-oi.R
\name{plot.pacea_oi}
\alias{plot.pacea_oi}
\title{Plot OISST data}
\usage{
\method{plot}{pacea_oi}(
  x,
  weeks.plot,
  months.plot,
  years.plot,
  bc = TRUE,
  eez = TRUE,
  restrict_plot = FALSE,
  buffer_proportion = 0.2,
  x_axis_labels = NULL,
  y_axis_labels = NULL,
  ...
)
}
\arguments{
\item{x}{a OISST \code{pacea_oi} object, which is an \code{sf} object.}

\item{weeks.plot}{numeric vector to indicate which weeks to plot. Defaults to current week (or most recent) available.}

\item{months.plot}{character or numeric vector to indicate which months to plot (e.g. \code{c(1, 2)}, \code{c("April", "may")}, \code{c(1, "April")}). Defaults to current month (or most recent) available.}

\item{years.plot}{numeric vector to indicate which years to plot. Defaults to current year (or most recent) available.}

\item{bc}{logical. Should BC coastline layer be plotted?}

\item{eez}{logical. Should BC EEZ layer be plotted?}

\item{restrict_plot}{logical. Should the plot be restricted to the spatial
domain of the data? If TRUE, the plot extent matches the data bounds and
bc_coast and bc_eez layers will not expand the plot limits. Default is FALSE.}

\item{buffer_proportion}{numeric. Buffer proportion to add around the data
bounds when restrict_plot is TRUE. Default is 0.2 (20\% in each direction).}

\item{x_axis_labels}{numeric vector. X-axis breaks to label when restrict_plot
is TRUE. If NULL, automatically hides overlapping labels.}

\item{y_axis_labels}{numeric vector. Y-axis breaks to label when restrict_plot
is TRUE. If NULL, automatically hides overlapping labels.}

\item{...}{other arguments to be passed on, but not currently used (\code{?ggplot}
says the same thing); this should remove a R-CMD-check warning.}
}
\value{
plot of the spatial data to the current device (returns nothing)
}
\description{
Plot for NOAA OISST sf objects using \code{ggplot()}. A quick visualization of data, specifying month(s) and year(s). For more options and configurable plots see vignette.
}
\examples{
\dontrun{
pdata <- oisst_7day
plot(pdata)
}
}
