% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-oianom.R
\name{plot.pacea_oianom}
\alias{plot.pacea_oianom}
\title{Plot anomaly of OISST spatiotemporal data layer}
\usage{
\method{plot}{pacea_oianom}(
  x,
  weeks.plot,
  months.plot,
  years.plot,
  clim.dat,
  bc = TRUE,
  eez = TRUE,
  restrict_plot = FALSE,
  buffer_proportion = 0.2,
  x_axis_labels = NULL,
  y_axis_labels = NULL,
  ...
)
}
\arguments{
\item{x}{an OISST \code{pacea_oianom} object; output from using \code{calc_anom()} of \code{oisst_7day} or \code{oisst_month} data}

\item{weeks.plot}{weeks to plot. Defaults to current week (if available)}

\item{months.plot}{months to plot. Defaults to current month (if available)}

\item{years.plot}{years to plot. Defaults to current year (if available)}

\item{clim.dat}{climatology data, obtained from using \code{calc_clim()}. If used, contours of deviations from climatology mean will be plotted}

\item{bc}{logical. Should BC coastline layer be plotted? Can only be plotted with one plot layer.}

\item{eez}{logical. Should BC EEZ layer be plotted? Can only be plotted with one plot layer.}

\item{restrict_plot}{logical. Should the plot be restricted to the spatial
domain of the data? If TRUE, the plot extent matches the data bounds and
bc_coast and bc_eez layers will not expand the plot limits. Default is FALSE.}

\item{buffer_proportion}{numeric. Buffer proportion to add around the data
bounds when restrict_plot is TRUE. Default is 0.2 (20\% in each
direction). Note that locations may be based on the centres of the grid
cells, hence you may need to play with \code{buffer_proportion} depending on your plot.}

\item{x_axis_labels}{numeric vector. X-axis breaks to label when restrict_plot
is TRUE. If NULL, automatically hides overlapping labels. Allows fine control
over which longitude values are displayed.}

\item{y_axis_labels}{numeric vector. Y-axis breaks to label when restrict_plot
is TRUE. If NULL, automatically hides overlapping labels. Allows fine control
over which latitude values are displayed.}

\item{...}{other arguments to be passed on, but not currently used (\code{?ggplot}
says the same thing); this should remove a R-CMD-check warning.}
}
\value{
plot of the spatial data to the current device (returns nothing)
}
\description{
Gets called from \code{pacea_oisst_anomalies_list()}, which uses the newer format
of climatology and anomalies TODO. Ensure consistent.
}
\examples{
\dontrun{
anom_data <- calc_anom(oisst_7day)
plot(anom_data)
}
}
