% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea_t.R
\name{plot.pacea_t}
\alias{plot.pacea_t}
\title{Plot a pacea time series object}
\usage{
\method{plot}{pacea_t}(obj)
}
\arguments{
\item{obj}{a `pacea_t` object, which is a time series.}
}
\description{
<desc>
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
