% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea_t.R
\name{plot.red_blue}
\alias{plot.red_blue}
\title{Plot the red/blue style of anomaly plot; internal function called from `plot-pacea_t()`.}
\usage{
\method{plot}{red_blue}(obj_lub, value, xlab, ylab, type, y_tick, x_tick_extra_years, ...)
}
\arguments{
\item{obj_lub}{obj a `pacea_t` object, which is a time series, with a date
column that is the lubridate `date` class.}

\item{value}{see `plot.pacea_t()`}

\item{xlab}{see `plot.pacea_t()`}

\item{ylab}{see `plot.pacea_t()`}

\item{type}{see `plot.pacea_t()`}

\item{y_tick}{see `plot.pacea_t()`}

\item{x_tick_extra_years}{see `plot.pacea_t()`}

\item{...}{see `plot.pacea_t()`}
}
\value{
plot of time series
}
\description{
Adapted from
https://stackoverflow.com/questions/74902499/shading-below-line-graph-in-r/74903305#74903305
}
\examples{
\dontrun{
# see plot.pacea_t()
}
}
\author{
Andrew Edwards
}
