% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-buoy-sst.R
\name{plot_buoy_sst_multiple}
\alias{plot_buoy_sst_multiple}
\title{Plot buoy SST data for multiple stations; called by `plot.pacea_buoy_sst()`}
\usage{
plot_buoy_sst_multiple(obj, stn_id, years, year_highlight)
}
\arguments{
\item{obj}{TODO}

\item{stn_id}{TODO}

\item{years}{TODO}

\item{year_highlight}{TODO}
}
\description{
TODO Develop once done single station. See buoy.Rmd vignette for some of
Andrea's code, probably have to combine that with single station code (not
just loop round single station).
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
