% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-buoy-anomaly-single.R
\name{plot_pacea_buoy_anomaly_single}
\alias{plot_pacea_buoy_anomaly_single}
\title{Plot anomalies for a single buoy, showing months on the y-axis.
Gets called from \code{plot.pacea_buoy_anomaly_list()} if \code{stn_id_to_plot} argument for
that has length 1.}
\usage{
plot_pacea_buoy_anomaly_single(
  pacea_buoy_anomaly_list,
  stn_id_to_plot,
  months,
  main,
  xlab,
  ylab
)
}
\arguments{
\item{pacea_buoy_anomaly_list}{object of class \code{pacea_buoy_anomaly_list}
obtained from running \code{calculate_anomaly()} on buoy data.}

\item{stn_id_to_plot}{station ID to plot}

\item{main}{title for the plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
a ggplot object
}
\description{
Plot anomalies for a single buoy, showing months on the y-axis.
Gets called from \code{plot.pacea_buoy_anomaly_list()} if \code{stn_id_to_plot} argument for
that has length 1.
}
\examples{
\dontrun{
TODO
all_buoys_anomalies <- calculate_anomaly(buoy_sst,
                             climatology_time = "month")
single_buoy_plot <- plot_pacea_buoy_anomaly_single(all_buoys_anomalies,
                                                   stn_id_to_plot = "C46132")
single_buoy_plot
}
}
\author{
Andrew Edwards
}
