% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-index.R
\name{plot_red_blue_bar}
\alias{plot_red_blue_bar}
\title{Plot the red/blue style of anomaly plot as barplots without any smoothing; internal function called from \code{plot.pacea_index()}.}
\usage{
plot_red_blue_bar(obj_lub, value, xlab, ylab, type, ...)
}
\arguments{
\item{obj_lub}{obj a \code{pacea_index} object, which is a time series, with a date
column that is the lubridate \code{date} class.}

\item{value}{which column of \code{obj} to plot (for \code{fraser_river_mean} and
\code{fraser_river_peak} objects this gets automatically switched to the
"value" column, as there is no "anomaly" column).}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label, the default is an attribute of the \code{pacea_index}
object.}

\item{type}{usual argument for \code{plot()}}

\item{...}{optional arguments passed onto \code{plot()}. Note that the x-axis is
constructed using a lubridate \code{date} object, so \code{xlim} needs to be a
\code{date} object (see example).}
}
\value{
plot of time series
}
\description{
Adapted from \code{plot_red_blue()}.
}
\examples{
\dontrun{
# see plot.pacea_index()
}
}
\author{
Andrew Edwards
}
