% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-with-uncertainty.R
\name{plot_with_uncertainty}
\alias{plot_with_uncertainty}
\title{Plot medians with uncertainty bars, such as for stock assessment
results; internal function called from `plot.pacea_recruitment()`.}
\usage{
plot_with_uncertainty(
  obj_lub,
  value,
  xlab,
  ylab,
  y_tick,
  x_tick_extra_years,
  uncertainty_bar_col,
  y_max,
  ...
)
}
\arguments{
\item{obj_lub}{obj a `pacea_t` object, which is a time series, with a date
column that is the lubridate `date` class.}

\item{value}{which column of `obj` to plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label, the default is an attribute of the `pacea_index`
object. TODO Note that this isn't automated yet to refer to anomaly or absolute
values, though for each time series there is probably only one choice}

\item{y_tick}{increment for y-axis ticks}

\item{x_tick_extra_years}{number of extra years to expand around the range
of data for which to add annual tick marks}

\item{uncertainty_bar_col}{colour for uncertainty bars for certain types of
plot (e.g. estimated fish recruitment)}

\item{y_max}{maximum y value for certain types of plot (use this if you get
an error when specifying `ylim`)}

\item{...}{optional arguments passed onto `plot()`. Note that the x-axis is
constructed using a lubridate `date` object, so `xlim` needs to be a
`date` object (see example).}
}
\value{
plot of time series with median as circle and bars for uncertainty.
}
\description{
Adapted from `make.mcmc.recruitment.plot()` from Pacific Hake assessment.
}
\examples{
\dontrun{
# see plot.pacea_recruitment()
}
}
\author{
Andrew Edwards
}
