% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{point2rast}
\alias{point2rast}
\title{Points to raster nearest neighbour interpolation}
\usage{
point2rast(
  data,
  spatobj,
  loc = c("x", "y"),
  cellsize,
  nnmax = 4,
  as = c("SpatRast", "SpatVect")
)
}
\arguments{
\item{data}{Vector, dataframe, or spatial (class: 'Spatial' or 'sf') input data to use for interpolation}

\item{spatobj}{Spatial object used to create extent for output raster}

\item{loc}{Vector of coordinate variable names or matrix of coordinates}

\item{cellsize}{Output raster cell size}

\item{nnmax}{Nearest neighbour maximum neighbours used for interpolation}

\item{as}{Output as raster 'SpatRast' or vector 'SpatVect'}
}
\value{
A terra SpatRaster or SpatVector object
}
\description{
Points to raster nearest neighbour interpolation
}
\examples{
\dontrun{
dat <- data.frame(x = runif(5, 0, 10), y = runif(5, 0, 10), var = rnorm(5))
extent <- st_bbox(c(xmin = 0, ymin = 0, xmax= 10, ymax = 10),crs = NA)
output <- point2rast(dat, extent, loc = c("x","y"), cellsize = 0.5, nnmax = 2, as = "SpatRast")
output
plot(output)
}
}
\author{
Travis Tai
}
