% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_to_area.R
\name{restrict_to_area}
\alias{restrict_to_area}
\title{Restrict a pacea spatial or spatiotemporal object to a given area, retaining
the class of the original object}
\usage{
restrict_to_area(data, area)
}
\arguments{
\item{data}{sf object of class \code{pacea_oi} TODO list them all}

\item{area}{sf object of a polygon for which to restrict the spatial data
to. Restriction is based on the centres of the cells of the data being
inside the area (thanks Kelsey!).}
}
\value{
object of the same class as \code{data}, but with spatial values only
within the defined area
}
\description{
Restrict a pacea spatial or spatiotemporal object to a given area, retaining
the class of the original object
}
\examples{
\dontrun{
sst_sasquatch <- restrict_to_area(oisst_month,
                                  sasquatch)
plot(sst_sasquatch, years = 2015:2016, months = 1:12)   # Sasquatch in a marine heatwave)
}
}
\author{
Andrew Edwards
}
