% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roms_average.R
\name{roms_average}
\alias{roms_average}
\title{Average ROMS output over a geographic area to give a time series}
\usage{
roms_average(
  pacea_st_obj,
  area = list(matrix(c(-127.1506, -128.2331, -129.3492, -127.9167, -127.1847, -126.82,
    -127.1506, 49.85766, 49, 48.99991, 50.11915, 50.40183, 50.24466, 49.85766), ncol =
    2))
)
}
\arguments{
\item{pacea_st_obj}{class `pacea_st` object, such as from ROMS output
or satellite sea surface temperature.}

\item{area}{a `list` object of numeric matrices (likely just one matrix)
with points in rows (see example) as longitude (negative) and latitude, to
be an input to `sf::st_polygon()`. Default gives an example area, based on
statistical fishing area 126 TODO double check definition.}
}
\value{
tibble with columns year, month, and value, similar to the oceanic indices
  (`oni` etc. but with no anomaly column; anomalies can be defined by the
  user if desired).
}
\description{
Given some ROMS output and a geographic area, average the values (such as
sea surface temperature or bottom oxygen) over the area to give a monthly
time series.
}
\examples{
\dontrun{
roms_sst <- roms_surface_temperature()
avg <- roms_average(roms_sst)
plot.pacea_index(avg, value = "value", ylab = "Average ROMS SST over area", style = "plain")
}
}
\author{
Andrew Edwards
}
