% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts-get-data.R
\name{roms_bottom_oxygen}
\alias{roms_bottom_oxygen}
\alias{roms_bottom_ph}
\alias{roms_bottom_salinity}
\alias{roms_bottom_temperature}
\alias{roms_surface_oxygen}
\alias{roms_surface_ph}
\alias{roms_surface_salinity}
\alias{roms_surface_temperature}
\alias{roms_avg0to40m_oxygen}
\alias{roms_avg0to40m_ph}
\alias{roms_avg0to40m_salinity}
\alias{roms_avg0to40m_temperature}
\alias{roms_avg40to100m_oxygen}
\alias{roms_avg40to100m_ph}
\alias{roms_avg40to100m_salinity}
\alias{roms_avg40to100m_temperature}
\alias{roms_avg100mtoBot_oxygen}
\alias{roms_avg100mtoBot_ph}
\alias{roms_avg100mtoBot_salinity}
\alias{roms_avg100mtoBot_temperature}
\alias{roms_phytoplankton}
\alias{roms_primaryproduction}
\title{ROMS environmental variable data}
\format{
A simple features dataframe.
}
\usage{
roms_bottom_oxygen(update = FALSE, ask = interactive(), force = FALSE)

roms_bottom_ph(update = FALSE, ask = interactive(), force = FALSE)

roms_bottom_salinity(update = FALSE, ask = interactive(), force = FALSE)

roms_bottom_temperature(update = FALSE, ask = interactive(), force = FALSE)

roms_surface_oxygen(update = FALSE, ask = interactive(), force = FALSE)

roms_surface_ph(update = FALSE, ask = interactive(), force = FALSE)

roms_surface_salinity(update = FALSE, ask = interactive(), force = FALSE)

roms_surface_temperature(update = FALSE, ask = interactive(), force = FALSE)

roms_avg0to40m_oxygen(update = FALSE, ask = interactive(), force = FALSE)

roms_avg0to40m_ph(update = FALSE, ask = interactive(), force = FALSE)

roms_avg0to40m_salinity(update = FALSE, ask = interactive(), force = FALSE)

roms_avg0to40m_temperature(update = FALSE, ask = interactive(), force = FALSE)

roms_avg40to100m_oxygen(update = FALSE, ask = interactive(), force = FALSE)

roms_avg40to100m_ph(update = FALSE, ask = interactive(), force = FALSE)

roms_avg40to100m_salinity(update = FALSE, ask = interactive(), force = FALSE)

roms_avg40to100m_temperature(
  update = FALSE,
  ask = interactive(),
  force = FALSE
)

roms_avg100mtoBot_oxygen(update = FALSE, ask = interactive(), force = FALSE)

roms_avg100mtoBot_ph(update = FALSE, ask = interactive(), force = FALSE)

roms_avg100mtoBot_salinity(update = FALSE, ask = interactive(), force = FALSE)

roms_avg100mtoBot_temperature(
  update = FALSE,
  ask = interactive(),
  force = FALSE
)

roms_phytoplankton(update = FALSE, ask = interactive(), force = FALSE)

roms_primaryproduction(update = FALSE, ask = interactive(), force = FALSE)
}
\arguments{
\item{update}{Logical. Would you like to check for a newer version of the layer?}

\item{ask}{Logical. Should the user be asked before downloading the data to local cache? Defaults to the value of interactive().}

\item{force}{Logical. Should download of data be forced? Overrides `ask` argument if TRUE.}
}
\value{
`sf` data object requested.
}
\description{
Regional Ocean Modelling System (ROMS) data for the BC Pacific region.

ROMS data interpolated to a 2km x 2km inshore grid and a 6km x 6km offshore grid. 
Data are monthly averages from 1993-2019. Each column represents a distinct 'year_month' combination (e.g. 2010_4).
}
\details{
\describe{
  \code{roms_*depth*_oxygen()} Dissolved oxygen measured as mmol-oxygen m^-3\cr
  \code{roms_*depth*_salinity()} Salinity in ppt\cr
  \code{roms_*depth*_temperature()} Temperature in oC\cr
  \code{roms_*depth*_ph()} pH\cr
  \code{roms_phytoplankton()} Total phytoplankton biomass in mmol-nitrogen m^-2\cr
  \code{roms_primaryproduction()} Total primary production in gC m^-2 d^-1
}

NOTE:\cr
\code{*depth*} must be replaced by one of the following depth categories:

\describe{
  \code{bottom}{sea bottom}\cr
  \code{0to40}{average between 0m and 40m depth}\cr
  \code{40to100}{average between 40m and 100m depth}\cr
  \code{100tobot} {:  average between 100m depth and sea bottom}\cr
  \code{surface}{sea surface}
}
}
\examples{
\dontrun{
my_data <- roms_bottom_oxygen()
}
}
