% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{romseez_poly}
\alias{romseez_poly}
\alias{inshore_poly}
\alias{offshore_poly}
\title{Polygons masking layers for BC waters}
\format{
A simple features polygon dataframe.
\describe{
  \item{romseez_poly}{BC EEZ boundary witha 10km buffer and clipped with ROMs data output boundary}
  \item{inshore_poly}{inshore (continental shelf) region of romseez_poly}
  \item{offshore_poly}{offshore (deep seafloor) region of roms eez_poly}
}

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.
}
\source{
Generated from running `data-raw/grids/make-mask-layer.R`.
}
\usage{
romseez_poly

inshore_poly

offshore_poly
}
\description{
Simple features objects of ROMs data, inshore, and offshore regions of BC Exclusive Economic Zone (EEZ) waters.
}
\details{
Bathymetric classifications were used to derive a line to separate continental shelf (inshore) and deep sea regions. Includes a 10km buffer around the EEZ and limited by ROMs data coverage.

Source of bathymetric classifications:
GIS hub Pacific Marine Habitat Classes - obtained from: https://www.gis-hub.ca/dataset/marine-habitat-classes
}
\examples{
\dontrun{
romseez_poly
inshore_poly
offshore_poly
plot(romseez_poly)
plot(inshore_poly)
plot(offshore_poly)
}

}
\keyword{datasets}
