% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rot}
\alias{rot}
\title{Create rotation matrix for an angle. Not needed I think. Took from Create_BC_Partition_New.Rmd
 <desc>}
\usage{
rot(a = 318.5)
}
\arguments{
\item{a}{Angle of rotation}
}
\value{
not-sure
}
\description{
Create rotation matrix for an angle. Not needed I think. Took from Create_BC_Partition_New.Rmd
 <desc>
}
\author{
Andrew Edwards
}
