% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sasquatch}
\alias{sasquatch}
\title{Sasquatch sf object for testing area options in code}
\format{
A simple features polygon dataframe.
}
\source{
Generated from running `data-raw/areas/sasquatch.R"
}
\usage{
sasquatch
}
\description{
A simple features object of a sasquatch, maruading off the west coast of
Vancouver Island and in Queen Charlotte Sound.
}
\examples{
\dontrun{
sasquatch
plot(sasquatch)
p <- plot(oisst_month)
p + geom_sf(data = sasquatch,
            fill = NA,
            colour = "black",
            linewidth = 1)
# Extract SST within the sasquatch
sasquatch_sst <- oisst_month[sasquatch, ]
class(sasquatch_sst) <- class(oisst_month)    # will make a function for that,
   or an option in the plotting, at some point
plot(sasquatch_sst)    # full coastline, no sasquatch outline

pp <- plot(sasquatch_sst)
pp + geom_sf(data = sasquatch,
             fill = NA,
             colour = "black",
             linewidth = 1)
# Can see that, as Kelsey explained, it's the centres of the SST squares that
# determine whether the square is considered part of the area or not.
}

}
\keyword{datasets}
