% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tran}
\alias{tran}
\title{Create a transition calculation, used in make_grid (could just put on there
 as I think only used once). Took from Create_BC_Partition_New.Rmd
 <desc>}
\usage{
tran(geo, ang, center)
}
\arguments{
\item{geo}{TODO}

\item{ang}{TODO}

\item{center}{TODO}
}
\value{
TODO
}
\description{
Create a transition calculation, used in make_grid (could just put on there
 as I think only used once). Took from Create_BC_Partition_New.Rmd
 <desc>
}
\author{
Andrew Edwards
}
