% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zooplankton_sog}
\alias{zooplankton_sog}
\title{Zooplankton anomalies in the Strait of Georgia}
\format{
A tibble also of class `pacea_zooplankton` with the following
  columns, where `total_biomass` onwards represents the annual anomaly of
  the specified species group in \eqn{\mbox{g m}^{-2}}:
\describe{
  \item{year:}{year of the calculated anomalies}
  \item{number_samples:}{total number of samples used for the annual
anomalies; this matches the Total column in Table 3 of Perry et al. (2021),
up to 2013; from 2014-2018 the numbers here are higher due to further
samples being processed. Data from 2019 onwards were not available for the
Perry at al. (2021) study.}
  \item{volume_filtered:}{total volume of water filtered for all tows,
  \eqn{\mbox{m}^3}.}
  \item{total_biomass:}{Total biomass}
  \item{amphipods_gammarid:}{Gammarid amphipods}
  \item{amphipods_hyperiid:}{Hyperiid amphipods}
  \item{benthic_larvae:}{Benthic larvae}
  \item{calanoid_copepods_large:}{Calanoid copepods (large)}
  \item{calanoid_copepods_medium:}{Calanoid copepods (medium)}
  \item{calanoid_copepods_small:}{Calanoid copepods (small)}
  \item{cephalopoda:}{Cephalopoda}
  \item{chaetognatha:}{Chaetognatha}
  \item{cladocera:}{Cladocera}
  \item{ctenophora:}{Ctenophora}
  \item{euphausiids:}{Euphausiids}
  \item{fish:}{Larval fish}
  \item{larvacea:}{Larvacea}
  \item{medusae:}{Medusae}
  \item{mysids:}{Mysids}
  \item{natantia:}{Natantia}
  \item{non_calanoid_copeopods:}{Non-calanoid copeopods}
  \item{ostracoda:}{Ostracoda}
  \item{other:}{Other taxa}
  \item{pelagic_polychaeta:}{Pelagic polychaeta}
  \item{pteropods:}{Pteropods}
  \item{repantia:}{Repantia}
  \item{scyphozoa:}{Scyphozoa}
  \item{siphonophorae:}{Siphonophorae}
 }
}
\source{
Anomalies calculated and provided by Kelly Young, then wrangled and imported using
  `data-raw/zooplankton/zooplankton.R`
}
\usage{
zooplankton_sog
}
\description{
Biomass anomalies of zooplankton biomass from 1996 onwards for the deep
waters, defined as
bottom depths greater than 50 m, of the central and northern Strait of
Georgia (see map in Perry et al., 2021). These extend the
time series calculated by Perry
et al. (2021) which was up to 2018, and as updated annually by Kelly Young in DFO's
State of the Pacific Ocean Report (latest is Young et al., 2023);
values here extend another year.
}
\details{
Following description is adapted from Young et al. (2023).

Zooplankton samples have been collected at approximately 20 standardized
stations monthly from February to October since 2015, with historic (but
sporadic sampling effort) data going back to 1995. Biomass was calculated
from the deep trawls, with bottom depths greater than 50 m and vertical net
haul samples that covered over 70\% of the water column. Data were restricted
to the central and northern Strait of Georgia (averaged together)
as they have the most complete time series available.

For historical comparison, the seasonal variability in
the zooplankton data was removed by calculating a regional, log10-scale
biomass anomaly for selected species groups for a given year. A multi-year
(1996 to most recent year) average seasonal cycle (climatology) was
calculated as a baseline to compare monthly conditions during any single
year. Seasonal anomalies were then averaged within each year to give an annual
anomaly (for details see Perry et al., 2021).

See examples and zooplankton vignette.

Perry, R.I., Young, K., Galbraith, M., Chandler, P., Velez-Espino A.,
Baillie S. (2021). Zooplankton variability in the Strait of Georgia, Canada,
and relationships with the marine survivals of Chinook and Coho salmon. PLoS
ONE 16(1):e0245941. https://doi.org/10.1371/journal.pone.0245941

Young, K., Galbraith, M., Sastri, A., and Perry, R.I. (2023). Zooplankton
status and trends in the central and northern Strait of Georgia, 2022. Pages
215-218 of Boldt, J.L., Joyce, E., Tucker, S., and Gauthier,
S. (Eds.). 2023. State of the physical, biological and selected fishery
resources of Pacific Canadian marine ecosystems in 2022. Canadian Technical
Report of Fisheries and Aquatic Sciences. 3542: viii+312 p.
https://www.dfo-mpo.gc.ca/oceans/publications/soto-rceo/2022/pac-technical-report-rapport-technique-eng.html
}
\examples{
\dontrun{
# Also see zooplankton vignette
zooplankton_sog
plot(zooplankton_sog)    # Default is total_biomass, calls plot.pacea_zooplankton()
plot(zooplankton_sog, species_group = "cladocera")  # y-axis name is automated
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
