% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zooplankton_sog_axis_names}
\alias{zooplankton_sog_axis_names}
\title{Zooplankton anomalies in the Strait of Georgia -- long names for y-axis labels}
\format{
A tibble with columns:
\describe{
\item{species_group_name:}{the available values for \code{species_group_name},
as given by the columns of \code{zooplankton_sog} from \code{total_biomass} onwards}
\item{axis_name :}{an expression to give a correct y label for the default
plots
(not just the \code{species_group_name} shorthand),
including getting the units (with superscripts) correct}
\item{in_perry_fig_s1:}{logical whether or not the species group is in the
20 figures shown in Figures S1 of Perry et al. (2021); see the zooplankton
vignette.}
\item{heatmap_axis_name :}{an expression to give a correct y label for the
ecosystem summary heatmap-style plot of indices (so do not need units)}
}
}
\source{
Created, based on Kelly Young's descriptions, in
\code{data-raw/zooplankton/zooplankton.R}
}
\usage{
zooplankton_sog_axis_names
}
\description{
For automated plotting of columns of \code{zooplankton_sog},
\code{zooplankton_sog_axis_names} contains expressions that correctly format to
give the y-axis label. Gets automatically used by
\code{plot.zooplankton_index()}, which gets automatically called by
\code{plot(zooplankton_sog)}. Should not really be needed by user. To see the
expressions used for the axis names, type
\code{zooplankton_sog_axis_names$axis_name}. Then to adapt one, find the
appropriate one and adapt that as necessary for \code{ylab}, see examples and
zooplankton vignette.
}
\examples{
\dontrun{
# Not really needed by user, gets used to automate the y-axis name, eg:
plot(zooplankton_sog, species_group = "cladocera")
# To see them all:
zooplankton_sog_axis_names$axis_name
# Then adapt the relevant one and, enclosing it in `expression()`, use as an
#  argument for `ylab` when making your plot. For example, cladocera one is:
# paste(plain(Cladocera) * " " * plain(anomaly) * ", " * log[10] *
     " " * g * " " * m^-2)
# Can manually adapt that to do:
plot(zooplankton_sog, species_group = "cladocera",
  ylab = expression(plain(Anomaly) * " " * plain(of) * " " * plain(Cladocera) * ", " * log[10] *
     " " * g * " " * m^-2))
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
