% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-clim.R
\name{calc_clim}
\alias{calc_clim}
\title{Calculate climatology}
\usage{
calc_clim(
  data,
  clim_years = c(1991:2020),
  clim_time = "month",
  time_period_return = "all"
)
}
\arguments{
\item{data}{pacea data object: BCCM, OISST, buoy_sst data only}

\item{clim_years}{climatology period years}

\item{clim_time}{time units (e.g. month) to summarize climatologies}

\item{time_period_return}{vector of value(s) for the specific time units to estimate climatologies (e.g. '4' for week 4 or April). Set to equal 'all' for all time units.}
}
\value{
climatology of data
}
\description{
Function for calculating climatology of a pacea data object.
}
\details{
The functions \code{calc_clim} and \code{calc_anom} output the climatology of a specified time period and anomaly values relative to a climatological period, respectively. These functions can be used for data from BCCM (e.g. \code{bccm_surface_temperature()}), OISST (e.g. \code{oisst_7day}), and buoy SST (e.g. \code{buoy_sst}).
}
\examples{
\dontrun{
# bccm data
pdata <- bccm_surface_temperature()
clim_bccm <- calc_clim(pdata)
head(clim_bccm)

clim_bccm2 <- calc_clim(pdata, time_period_return = c(1, 6))
head(clim_bccm2)

# oisst data
clim_oi <- calc_clim(oisst_7day)
head(clim_oi)

# pacea buoy sst data
clim_buoy <- calc_clim(buoy_sst)
head(clim_buoy)

clim_buoy2 <- calc_clim(buot_sst, clim_time = "week")
}
}
