% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_climatology_anomaly.R
\name{calc_clim}
\alias{calc_clim}
\alias{calc_anomaly}
\title{Calculate climatology and anomaly}
\usage{
calc_clim(data, clim_years = c(1991:2020), temporal_FUN = "month")

calc_anomaly(data, clim_years = c(1991:2020), temporal_FUN = "month")
}
\arguments{
\item{data}{}

\item{clim_years}{}

\item{temporal_FUN}{}
}
\value{
anomaly data values of same class as input pacea data

anomaly data values of same class as input pacea data
}
\description{
Function for calculating climatology or anomaly of a pacea data object.
}
\details{
The following data objects in 'pacea' are in the absolute values of the variable: BCCM, OISST, and buoy SST data. The functions `calc_clim` and `calc_anom` output the climatology of a specified time period and anomaly values relative to a climtological period, respectively.
}
\examples{
\dontrun{
# bccm data
pdata <- bccm_surface_temperature()
clim_bccm <- calc_clim(pdata)
head(clim_bccm)
plot(clim_bccm)

# oisst data
clim_oi <- calc_clim(oisst_7day)
head(clim_oi)
plot(clim_oi)


# pacea buoy sst data
clim_buoy <- calc_clim(buoy_sst)
head(clim_buoy)
plot(clim_buoy)
}
\dontrun{
}
}
