% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-anomaly-pacea-buoy.R
\name{calculate_anomalies.pacea_buoy}
\alias{calculate_anomalies.pacea_buoy}
\title{Calculate climatology and anomalies ANDY DOING JUST buoy_sst object adapting from Travis's,
want to tailor the options for each pacea object, so use generics. TODO
putting anomaly calcs in here also, then return a list object that has
both. So change name to anomaly at some point.}
\usage{
\method{calculate_anomalies}{pacea_buoy}(
  data,
  climatology_years = 1991:2020,
  climatology_time = "month",
  time_period_return = "all",
  years_return = NULL,
  min_days_per_month = 15,
  max_consecutive_missing_days = 6
)
}
\arguments{
\item{data}{\code{buoy_sst} pacea data object}

\item{climatology_years}{climatology period years}

\item{climatology_time}{time units ("month" or "week") to summarize
climatologies and anomalies, but only "month" is currently available (there
is some functionality to use "week" also -- contact Andy if this would be useful).}

\item{time_period_return}{vector of value(s) for the specific time units to estimate climatologies (e.g. '4' for week 4 or April). Set to equal 'all' for all time units.}

\item{years_return}{vector of value(s) to return the years of
interest. Defaults to all years in input data}

\item{min_days_per_month}{minimum number of daily SST values required in a
time period (month or week) for that period to be included in climatology
calculation and anomalies calculation. Defaults to 15 days per month.}

\item{max_consecutive_missing_days}{maximum number of consecutive days allowed
to be missing (NA) within a time period (month or week). If a time period has
more than this many consecutive NAs, it is excluded from climatology and anomalies
calculation. Defaults to 6 days.}
}
\value{
TODO list object of climatology of data and anomalies, of class
\code{pacea_buoy_anomalies_list}. Note that \code{climatology_years} will be the
prescribed years, but these may not be available for all buoys.
}
\description{
Function for calculating climatology of a \code{buoy_sst} pacea data object.
}
\details{
TODO The functions \code{calc_clim} and \code{calc_anom} output the climatology of a specified time period and anomaly values relative to a climtological period, respectively. These functions can be used for data from BCCM (e.g. \code{bccm_surface_temperature()}), OISST (e.g. \code{oisst_7day}), and buoy SST (e.g. \code{buoy_sst}).
}
\examples{
\dontrun{
# Will integrate options into function better TODO see vignette
one_stn_id_example <- "C46146"
buoy_example <- buoy_sst \%>\%
  filter(stn_id == one_stn_id_example)
res <- calculate_anomalies(buoy_example,
                         climatology_time = "month")
}
}
