% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-anomalies-pacea-oi.R,
%   R/calculate-anomalies.R
\name{calculate_anomalies.pacea_oi}
\alias{calculate_anomalies.pacea_oi}
\alias{calculate_anomalies}
\title{TODO all the help Calculate climatology and anomalies for oisst data.}
\usage{
\method{calculate_anomalies}{pacea_oi}(
  data,
  climatology_years = 1991:2020,
  climatology_time = "month",
  time_period_return = "all",
  years_return = NULL,
  area = NULL
)

calculate_anomalies(data, ...)
}
\arguments{
\item{data}{One of:
\itemize{
\item \code{size_spectrum_mlebins} object
\item \code{determine_xmin_and_fit_mlebins} object
\item \code{remove_outliers_mlebins} object
}}

\item{climatology_years}{climatology period years}

\item{climatology_time}{time units ("month" or "week") to summarize
climatologies and anomalies, but only "month" is currently available (there
is some functionality to use "week" also -- contact Andy if this would be useful).}

\item{time_period_return}{vector of value(s) for the specific time units to estimate climatologies (e.g. '4' for week 4 or April). Set to equal 'all' for all time units.}

\item{years_return}{vector of value(s) to return the years of
interest. Defaults to all years in input data}

\item{area}{characters string of a named area (only \code{pfma_area_126} at the
moment, which is harwired into the function and will be removed), or
data.frame, tibble, or matrix with columns as latitude and longitude of
points around the region, with the final point equalling the first one (so
defining a closed  polygon). If NULL then the whole region of data. TODO NOT
ALL IMPLEMENTED YET}

\item{...}{arguments passed onto the respective \verb{detect_outliers.<class>} function}
}
\value{
pacea_buoy_anomalies_list object (also a list) of climatology of data and anomalies, of class
\code{pacea_buoy_anomalies_list}. Note that \code{climatology_years} will be the
prescribed years, but these may not be available for all buoys.

tibble containing just the data values needed for
calculations, which are the original four columns:
\itemize{
\item \code{species}
\item \code{bin_min}
\item \code{bin_max}
\item \code{bin_count}
plus
\item \code{gap} which is \code{bin_min} minus the previous \code{bin_min}
\item \code{gap_ratio} which is the ratio of \code{gap} to the second largest gap, so makes
it easy to see if the largest gap is much much larger than the second largest.
}
}
\description{
Function for calculating climatology and subsequent anomalies of a \code{buoy_sst} pacea data object.
}
\examples{
\dontrun{
area_126 <- tibble(
  lat = c(-127.1506, -128.2331, -129.3492, -127.9167, -127.1847, -126.8200,
-127.1506),
  lon = c(49.85766, 49.00000, 48.99991, 50.11915, 50.40183, 50.24466, 49.85766))
res <- calculate_anomalies(oisst_month, area = area_126)
}
\dontrun{
# See ?remove_outliers for link to example code
}
}
\author{
Andrew Edwards and Travis Tai

Andrew Edwards
}
