% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-anomaly-pacea-buoy.R
\name{calculate_anomaly.pacea_buoy}
\alias{calculate_anomaly.pacea_buoy}
\title{Calculate climatology and anomaly ANDY DOING JUST buoy_sst object adapting from Travis's,
want to tailor the options for each pacea object, so use generics. TODO
putting anomaly calcs in here also, then return a list object that has
both. So change name to anomaly at some point.}
\usage{
\method{calculate_anomaly}{pacea_buoy}(
  data,
  climatology_years = c(1991:2020),
  climatology_time = "month",
  time_period_return = "all",
  years_return = NULL
)
}
\arguments{
\item{data}{\code{buoy_sst} pacea data object}

\item{climatology_years}{climatology period years}

\item{climatology_time}{time units (e.g. month) to summarize climatologies}

\item{time_period_return}{vector of value(s) for the specific time units to estimate climatologies (e.g. '4' for week 4 or April). Set to equal 'all' for all time units.}

\item{years_return}{vector of value(s) to return the years of
interest. Defaults to all years in input data}
}
\value{
TODO list object of climatology of data and anomaly, of class \code{pacea_buoy_anomaly_list}
}
\description{
Function for calculating climatology of a buoy_sstpacea data object.
}
\details{
TODO The functions \code{calc_clim} and \code{calc_anom} output the climatology of a specified time period and anomaly values relative to a climtological period, respectively. These functions can be used for data from BCCM (e.g. \code{bccm_surface_temperature()}), OISST (e.g. \code{oisst_7day}), and buoy SST (e.g. \code{buoy_sst}).
}
\examples{
\dontrun{
# Will integrate options into function better TODO
one_stn_id_example <- "C46146"
buoy_example <- buoy_sst \%>\%
  filter(stn_id == one_stn_id_example)
res <- calculate_anomaly(buoy_example,
                         climatology_time = "month")
}
}
