% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-anomaly.R
\name{calculate_anomaly}
\alias{calculate_anomaly}
\title{TODO this is for \code{calculate_climatology()}. Edit whole help file.
If already run through \code{remove_outliers.size_spectrum_mlebins()} then it wil
use \code{detect_outliers.remove_outliers.mlebins()} and can keep iterating
that. See \link{remove_outliers} for further details and example code.}
\usage{
calculate_anomaly(data, ...)
}
\arguments{
\item{...}{arguments passed onto the respective \verb{detect_outliers.<class>} function}

\item{res}{One of:
\itemize{
\item \code{size_spectrum_mlebins} object
\item \code{determine_xmin_and_fit_mlebins} object
\item \code{remove_outliers_mlebins} object
}}
}
\value{
tibble containing just the data values needed for
calculations, which are the original four columns:
\itemize{
\item \code{species}
\item \code{bin_min}
\item \code{bin_max}
\item \code{bin_count}
plus
\item \code{gap} which is \code{bin_min} minus the previous \code{bin_min}
\item \code{gap_ratio} which is the ratio of \code{gap} to the second largest gap, so makes
it easy to see if the largest gap is much much larger than the second largest.
}
}
\description{
TODO this is for \code{calculate_climatology()}. Edit whole help file.
If already run through \code{remove_outliers.size_spectrum_mlebins()} then it wil
use \code{detect_outliers.remove_outliers.mlebins()} and can keep iterating
that. See \link{remove_outliers} for further details and example code.
}
\examples{
\dontrun{
# See ?remove_outliers for link to example code
}
}
\author{
Andrew Edwards
}
