% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-index-changed.R
\name{check_index_changed}
\alias{check_index_changed}
\title{Check if a newly downloaded index has changed from one currently in pacea}
\usage{
check_index_changed(old, new)
}
\arguments{
\item{old}{original index (`pacea_t` object) saved in pacea}

\item{new}{newly downloaded and wrangled index (`pacea_t` object)}
}
\value{
logical if the index has changed (and we would then want to resave it)
}
\description{
If index has changed from the saved one then save the new version.
}
\examples{
\dontrun{
check_index_changed(oni, oni[-10, ]))
}
}
\author{
Andrew Edwards
}
