% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-herring-object.R
\name{create_herring_object}
\alias{create_herring_object}
\title{Create herring object to then save as data (five regions so best to functionalise).}
\usage{
create_herring_object(assess_yr, region, herring_dir = NULL)
}
\arguments{
\item{assess_yr}{The year of the assessment (yielding stock status in that year).}
}
\value{
List of recruitment and biomass objects, to then be saved in package.
}
\description{
Used in `data-raw/herring/herring.R` to create data objects; not really
needed by `pacea` users.
}
\examples{
\dontrun{
}
}
\author{
Andrew Edwards
}
