% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-index-pacea-biomass-herring.R,
%   R/create-index-pacea-biomass.R, R/create-index-pacea-buoy.R,
%   R/create-index-pacea-harbour-seals.R, R/create-index-pacea-oi.R,
%   R/create-index-pacea-oisst-anomalies-list.R,
%   R/create-index-pacea-recruitment-herring.R,
%   R/create-index-pacea-recruitment.R, R/create-index-pacea-zooplankton.R,
%   R/create-index.R
\name{create_index.pacea_biomass_herring}
\alias{create_index.pacea_biomass_herring}
\alias{create_index.pacea_biomass}
\alias{create_index.pacea_buoy}
\alias{create_index.pacea_harbour_seals}
\alias{create_index.pacea_oi}
\alias{create_index.pacea_oisst_anomalies_list}
\alias{create_index.pacea_recruitment_herring}
\alias{create_index.pacea_recruitment}
\alias{create_index.pacea_zooplankton}
\alias{create_index}
\title{Create an annual standardised index based on a pacea object}
\usage{
\method{create_index}{pacea_biomass_herring}(
  data,
  years = NULL,
  herring_region = "WCVI",
  index_label = NULL,
  index_name = NULL
)

\method{create_index}{pacea_biomass}(data, years = NULL, index_label = "Hake spawning biomass")

\method{create_index}{pacea_buoy}(
  data,
  years = NULL,
  stn_id = "C46205",
  index_label = NULL,
  index_name = NULL,
  index_statistic = "anomalies",
  require_requested_months = NULL,
  ...
)

\method{create_index}{pacea_harbour_seals}(
  data,
  years = NULL,
  seal_region = "WCVI",
  index_label = NULL,
  index_name = NULL,
  index_statistic = "mean"
)

\method{create_index}{pacea_oi}(
  data,
  years = NULL,
  index_label = NULL,
  index_name = NULL,
  index_statistic = "anomalies",
  area = NULL,
  months = NULL,
  require_requested_months = NULL,
  ...
)

\method{create_index}{pacea_oisst_anomalies_list}(
  data,
  years = NULL,
  index_label = NULL,
  index_name = NULL,
  index_statistic = "anomalies",
  area = NULL,
  require_requested_months = NULL,
  ...
)

\method{create_index}{pacea_recruitment_herring}(
  data,
  years = NULL,
  herring_region = "WCVI",
  index_label = NULL,
  index_name = NULL,
  index_statistic = "median"
)

\method{create_index}{pacea_recruitment}(
  data,
  years = NULL,
  index_label = "Hake age-0 recruitment",
  index_name = NULL,
  index_statistic = "median"
)

\method{create_index}{pacea_zooplankton}(
  data,
  years = NULL,
  zooplankton_species_group = "total_biomass",
  index_label = NULL,
  index_name = NULL,
  index_statistic = "anomaly"
)

create_index(data, ...)
}
\arguments{
\item{data}{One of the data objects (or a reduced version that retains the
class of the object) in pacea:
Biological populations:
\itemize{
\item \code{herring_recruitment}
\item \code{herring_spawning_biomass}
\item \code{hake_recruitment}
\item \code{hake_biomass}
\item \code{harbour_seals}
\item \code{zooplankton_sog} (zooplankton data in the Strait of Georgia)
Oceanography:
\item \code{buoy_sst} - sea-surface temperature from buoys
\item \code{oisst_month} - optimally interpolated sea-surface temperature
}}

\item{years}{numeric vector years to restrict the data to, and then to
normalise over. If \code{NULL}, the default, then use the full range available
for the type of data being used. For \code{buoy_sst} this will be the available
data for the chosen \code{stn_id} buoy. TODO check what happens for other types
of data.}

\item{herring_region}{string for the region to be used when making an index related
to \code{herring_recruitment} or \code{herring_spawning_biomass}.}

\item{index_label}{character string for the name to be used as the label in
the plotting. Default is for hake recruitment, and will need to be manually
changed if doing another type of recruitment (except for herring, which is
automated; currently there are no other recruitments in pacea).}

\item{index_name}{character string for the name of the index. If \code{NULL} (default),
the name is derived from the variable name of \code{data}. Used internally when delegating
between \code{create_index} methods to preserve the original variable name.}

\item{stn_id}{string of the station ID (\code{stn_id} value OR
\code{name} value from \code{buoy_metadata}) to create an index from the \code{buoy_sst} data
object.}

\item{index_statistic}{string for the column of \code{data} to use to create the
index. Default is \code{median} (for hake recruitment), but user can choose any
column of the data. When looking at harbour seals the function automatically
uses the \code{mean}. For \code{buoy_sst} it is either \code{anomalies} or
\code{mean}, to prescribe whether to create the index based on means of anomalies
(over the prescribed months) from a climatology, or just the means of the
actual SST values. These may well be the same anyway, or only differ if no
climatology can be built (which can depend on the years used for the
climatology). For \code{oisst_month} only \code{anomalies} (the default) can be used.}

\item{require_requested_months}{number of months that must be available
out of the requested \code{months} to compute an average anomaly to then use to
create the index. If \code{NULL} (default), defaults to \code{length(months)}. See
\code{\link[=calculate_anomalies.pacea_buoy]{calculate_anomalies.pacea_buoy()}} also for more options for \code{buoy_sst} data, such as the
minimum daily SST values required in a month to do the calculations.
(i.e. every month requested needs enough daily values).}

\item{...}{other arguments passed onto the respective method, and also onto
\code{\link[=calculate_anomalies.pacea_buoy]{calculate_anomalies.pacea_buoy()}} for \code{buoy_sst} data. TODO test these}

\item{seal_region}{string for the region to be used when making an index related
to \code{harbour_seals}.}

\item{area}{characters string of a named area for creating an index as an
average of a spatial area, only for OISST at the moment TODO update. Also (only \code{pfma_area_126} at the
moment, which is hardwired into the function and will be removed), or
data.frame, tibble, or matrix with columns as latitude and longitude of
points around the region, with the final point equalling the first one (so
defining a closed  polygon). If NULL then the whole region of data. TODO NOT
ALL IMPLEMENTED YET}

\item{months}{vector of months (default \code{4} for April) for \code{buoy_sst} and
\code{oisst_month}
data, over which to average the values to create the annual index. See \code{\link[=plot.pacea_buoy_anomalies_list]{plot.pacea_buoy_anomalies_list()}} for full details of options
(e.g. \code{c(11, 12, 1, 2, 3)} for a winter average). TODO does not seem to be
used at the moment, according to \code{check_man()}, 24/9/26.}

\item{zooplankton_species_group}{string for the zooplankton species group
to be used from \code{zooplankton_sog}, must be one of the biological columns
(\code{total_biomass} onwards) of \code{zooplankton_sog}.}
}
\value{
tibble of class \code{pacea_standardised_index} containing columns of
\itemize{
\item \code{index} - name of the index
\item \code{year} - year of the value
\item \code{value} - the value of the index in that year
\item \code{index_label} - label to be used for plotting
}
}
\description{
Create an annual index based on a pacea object. The annual value can be
calculated from annual values, such as \code{hake_recruitment}, or averaged for
specific monthly values (e.g. PDO for Apr to Sep). See the examples and
vignette TODO
Calculations are based on the class of the object: \code{\link[=create_index]{create_index()}} is a generic function that calls \verb{create_index.<class>()}
where \verb{<class>} is \code{class(data)}. \code{\link[=create_index.pacea_recruitment]{create_index.pacea_recruitment()}} is a
kind of master file that gets used by others when possible (could not for
herring), and just needs columns \code{year} and a column named
\code{index_statistic}, other information is supplied from other functions in the
arguments.
}
\details{
\code{\link[=create_index.pacea_harbour_seals]{create_index.pacea_harbour_seals()}} linearly interpolates the
roughly eight-monthly population estimates to the 1st January to give a
single value for each year. (Hake assessment outputs are also for the start
of the year, TODO herring I think might be 1st April).

\code{\link[=create_index.pacea_oi]{create_index.pacea_oi()}} works directly on the \code{oisst_month} data,
by using \code{\link[=calculate_anomalies.pacea_oi]{calculate_anomalies.pacea_oi()}} and then calculating a mean
anomaly averaged over the specified \code{months} specified and spatial \code{area} (one value for
each year)

\code{\link[=create_index.pacea_oisst_anomalies_list]{create_index.pacea_oisst_anomalies_list()}} works on the results of
\code{\link[=calculate_anomalies]{calculate_anomalies()}} on \code{oisst_month} data, which can be used before
calculating the index  to show spatial variation. The resulting
\code{\link[=create_index]{create_index()}} on the results then
creates an annual index of the mean anomaly averaged over the specified \code{months}
and spatial \code{area} (one value for  each year).

TODO add a test to make sure they match, and default months is always April.

The index is normalised (subtract the mean and divide by the standard
deviation) across the specified years. This means that different types of
indices can be compared to each other. Indices are returned in a
standardised format for plotting. TODO: Medians are used if there is uncertainties
are given in the original data, though functions may have choices
}
\examples{
\dontrun{
create_index(hake_recruitment)
}
}
\author{
Andrew Edwards
}
