% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-index-pacea-biomass-herring.R,
%   R/create-index-pacea-biomass.R, R/create-index-pacea-harbour-seals.R,
%   R/create-index-pacea-recruitment-herring.R,
%   R/create-index-pacea-recruitment.R, R/create-index.R
\name{create_index.pacea_biomass_herring}
\alias{create_index.pacea_biomass_herring}
\alias{create_index.pacea_biomass}
\alias{create_index.pacea_harbour_seals}
\alias{create_index.pacea_recruitment_herring}
\alias{create_index.pacea_recruitment}
\alias{create_index}
\title{Create an annual standardised index based on a pacea object}
\usage{
\method{create_index}{pacea_biomass_herring}(
  data,
  years = NULL,
  herring_region = "WCVI",
  index_label = NULL,
  index_name = NULL
)

\method{create_index}{pacea_biomass}(data, years = NULL, index_label = "Hake spawning biomass")

\method{create_index}{pacea_harbour_seals}(
  data,
  years = NULL,
  region = "WCVI",
  index_label = NULL,
  index_name = NULL,
  index_statistic = "mean"
)

\method{create_index}{pacea_recruitment_herring}(
  data,
  years = NULL,
  herring_region = "WCVI",
  index_label = NULL,
  index_name = NULL
)

\method{create_index}{pacea_recruitment}(
  data,
  years = NULL,
  index_label = "Hake age-0 recruitment",
  index_name = NULL,
  index_statistic = "median"
)

create_index(data, ...)
}
\arguments{
\item{data}{One of: TODO
\itemize{
\item \code{hake_recruitment} (of class \code{pacea_recruitment}).
}}

\item{years}{numeric vector years to restrict the data to, and then to normalise over.}

\item{herring_region}{string for the region to be used when making an index related
to \code{herring_recruitment} or TODO}

\item{index_label}{character string for the name to be used as the label in
the plotting. Default is for hake recruitment, and will need to be manually
changed if doing another type of recruitment (except for herring, which is
automated; currently there are no other recruitments in pacea).}

\item{index_name}{character string for the name of the index. If \code{NULL} (default),
the name is derived from the variable name of \code{data}. Used internally when delegating
between \code{create_index} methods to preserve the original variable name.}

\item{index_statistic}{string for the column of \code{data} to use to create the
index. Default is \code{median} (for hake recruitment), but user can choose any
column of the data. When looking at harbour seals the function automatically
uses the \code{mean}.}

\item{...}{arguments passed onto the respective \verb{create_index.<class>}
function, such as:}
}
\value{
tibble of class \code{pacea_standardised_index} containing columns of
\itemize{
\item \code{index} - name of the index
\item \code{year} - year of the value
\item \code{value} - the value of the index in that year
\item \code{index_label} - label to be used for plotting
}
}
\description{
Create an annual index based on a pacea object. The annual value can be
calculated from annual values, such as \code{hake_recruitment}, or averaged for
specific monthly values (e.g. PDO for Apr to Sep). See the examples and
vignette TODO
Calculations are based on the class of the object: \code{\link[=create_index]{create_index()}} is a generic function that calls \verb{create_index.<class>()}
where \verb{<class>} is \code{class(data)}. \code{\link[=create-index.pacea_recruitment]{create-index.pacea_recruitment()}} is a
kind of master file that gets used by others when possible (could not for herring).
}
\details{
The index is normalised (subtract the mean and divide by the standard
deviation) across the specified years. This means that different types of
indices can be compared to each other. Indices are returned in a
standardised format for plotting. TODO: Medians are used if there is uncertainties
are given in the original data, though functions may have choices
}
\examples{
\dontrun{
create_index(hake_recruitment)
}
}
\author{
Andrew Edwards
}
