% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-index-pacea-recruitment.R,
%   R/create-index.R
\name{create_index.pacea_recruitment}
\alias{create_index.pacea_recruitment}
\alias{create_index}
\title{Create an index time series based on hake recruitment estimates.}
\usage{
\method{create_index}{pacea_recruitment}(data, years = NULL, index_label = "Hake recruitment", ...)

create_index(data, ...)
}
\arguments{
\item{data}{One of: TODO
\itemize{
\item \code{size_spectrum_mlebins} object
\item \code{determine_xmin_and_fit_mlebins} object
\item \code{remove_outliers_mlebins} object
}}

\item{...}{arguments passed onto the respective \verb{detect_outliers.<class>} function}
}
\value{
tibble containing columns of
\itemize{
\item \code{index} - name of the index
\item \code{year} - year of the value
\item \code{index_plot_value} - the value of the index in that year
}

tibble containing columns of
\itemize{
\item \code{index} - name of the index
\item \code{year} - year of the value
\item \code{index_plot_value} - the value of the index in that year
}
}
\description{
TODO put all the helps into create_index() when done.

Create an annual index based on a pacea object. The annual value can be
tailored to be anything. See the examples and vignette TODO
Calculations are based on the class of the object.
}
\details{
pacea_recruitment is just for hake, should add hake at the end at some
point maybe. TODO
Create an annual index based on estimates of \code{hake_recruitment} saved in
pacea.

years identify the vector of years to restrict to, and then to normalise over.

The index is normalised (subtract the mean and divide by the standard
deviation) across the specified years. This means that different types of
indices can be compared to each other. Indices are returned in a
standardised format for plotting. TODO: Medians are used if there is uncertainties
are given in the original data, though functions may have choices
}
\examples{
\dontrun{
create_index(hake_recruitment)
}
\dontrun{
# See ?remove_outliers for link to example code
}
}
\author{
Andrew Edwards
}
