% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-zenodo-data.R
\name{get_zenodo_data}
\alias{get_zenodo_data}
\title{Get large data products from Zenodo (if the file is not already cached locally)}
\usage{
get_zenodo_data(
  layer,
  update = FALSE,
  ask = interactive(),
  force = FALSE,
  version = "01",
  cache_subfolder,
  zenodo_doi = "10.5281/zenodo.14019141",
  timeout_value = 7200
)
}
\arguments{
\item{layer}{Name of the data object.}

\item{update}{Logical. Would you like to check for a newer version of the layer?}

\item{ask}{Logical. Should the user be asked before downloading the data to local cache? Defaults to the value of interactive().}

\item{force}{Logical. Should download of data be forced? Overrides `ask`
argument if TRUE. Does not force a download if the file is already cached locally.}

\item{version}{Character. Version number of data on Zenodo (as in what is
appended at the end of the filenames).}

\item{cache_subfolder}{Character. The folder
`paste0(pacea::pacea_cache(), cache_subfolder)` will be
searched first for the desired model output file, and if the no file is
there then it will be downloaded from Zenodo into this directory. Is set to
`hotssea` but gets automatically changed to `bccm_full` if `layer`
contains `bccm` and `full`.}

\item{zenodo_doi}{character. The DOI of the zenodo record to download files
from. Default is for `hotssea` but gets automatically changed to the DOI
for `bccm_full` if `layer` contains `bccm` and `full`.}

\item{timeout_value}{numeric Timeout (seconds) for downloading a file from
the internet. Gets used in [zen4R::download_zenodo()].}
}
\value{
Data object requested
}
\description{
HOTSSea and full BCCM model results have been wrangled from the original
 results and are hosted on Zenodo due to large filesizes. This function
 downloads an individual file
 and puts it in the appropriate cache on a user's local computer. Users
 should use the functions that
 call this, see [hotssea_all_variables()] and [bccm_all_variables_full()],
 and also the help for invidual functions such as
 [hotssea_surface_salinity_min()] and [bccm_bottom_temperature_full()].
}
\details{
Based on [get_bccm_data()], but getting from Zenodo not `pacea-data/`.
}
\examples{
\dontrun{
h <- get_zenodo_data("hotssea_surface_temperature_max")
plot(h)

b <- get_zenodo_data("bccm_bottom_oxygen_full")   # Only took 20 seconds on
  home network, for a 108Mb file.
plot(b)
}

}
\author{
Andrew Edwards and Travis Tai
}
