% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hake_biomass}
\alias{hake_biomass}
\title{Pacific Hake annual spawning stock biomass (females only) as estimated by the 2023 stock
assessment.}
\format{
A tibble also of class `pacea_biomass` with columns:
\describe{
  \item{year:}{year of the estimate of spawning stock biomass (mature
  females), where the estimate is for the start of the year}
  \item{low:}{low end (2.5th percentile) of the 95\% credible interval for
  biomass, units as for `median`}
  \item{median:}{median estimate of biomass, in millions of tonnes of female
  spawning biomass at the start of the year}
  \item{low:}{high end (97.5th percentile) of the 95\% credible interval for
  biomass, units as for `median`}
 }
}
\source{
Generated from Andy running (in the hake repository)
  `hake-assessment/sandbox/andy/pacea-save/pacea-save.R` and then here
  `data-raw/groundfish/hake.R`.
}
\usage{
hake_biomass
}
\description{
The Pacific Hake stock is managed and assessed through an Agreement between
Canada and the United States. The spawning stock biomass (mature females) estimates come from the most
recent joint stock assessment, and, importantly, are for the coastwide stock
from California to British Columbia.
}
\details{
See `?hake_recruitment` for further details and reference.
}
\examples{
\dontrun{
hake_biomass
plot(hake_biomass)
}
}
\author{
Andrew Edwards
}
\keyword{datasets}
