% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oisst_7day}
\alias{oisst_7day}
\alias{oisst_7dayclim}
\alias{oisst_month}
\alias{oisst_monthclim}
\title{Optimal interpolation SST}
\format{
`oisst_7day`
A object of class sf with 2051650 rows and 8 columns.

`oisst_7dayclim`
An object of class sf with 49025 rows and 6 columns.

`oisst_month`
An object of class sf with 464350 rows and 8 columns.

`oisst_monthclim`
An object of class sf with 11100 rows and 6 columns.
}
\source{
https://www.ncei.noaa.gov/products/optimum-interpolation-sst
Generated from running `data-raw/erddap/OISST.R`.
}
\usage{
oisst_7day

oisst_7dayclim

oisst_month

oisst_monthclim
}
\description{
Simple features objects of NOAA's Optimal Interpolation SST
}
\details{
From NOAA:
"The NOAA 1/4° Daily Optimum Interpolation Sea Surface Temperature (OISST) is a long term Climate Data Record that incorporates observations from different platforms (satellites, ships, buoys and Argo floats) into a regular global grid. The dataset is interpolated to fill gaps on the grid and create a spatially complete map of sea surface temperature. Satellite and ship observations are referenced to buoys to compensate for platform differences and sensor biases."

The data starts from 1981-09-01 and ends in the previous month from the current month. These data will be updated monthly. 

\describe{
  \item{oisst_7day}{7-day mean sea surface temperature (degrees Celsius)}
  \item{oisst_7dayclim}{7-day mean climatology (1991-2020) for sea surface temperature (degrees Celsius)}
  \item{oisst_month}{Monthly mean sea surface temperature in degrees Celsius}
  \item{oisst_monthclim}{Monthly climatology (1991-2020) for sea surface temperature (degrees Celsius)}
}
}
\examples{
\dontrun{
head(oisst_7day)
head(oisst_7dayclim)
}

}
\keyword{datasets}
