% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Restrict two time series to only the overlapping dates (only years for now)}
\usage{
overlap(obj_1, obj_2, region = "SOG")
}
\arguments{
\item{obj_1}{data.frame first time series}

\item{obj_2}{data.frame second time series}

\item{region}{character corresponding the region to be used for
herring-related objects (of class \code{pacea_recruitment_herring} or
\code{pacea_biomass_herring})}
}
\value{
tibble with first column \code{year} and the rest being the columns of
\code{obj_1} and \code{obj_2} restricted to the overlapping years
}
\description{
TODO Extend to all the classes of pacea objects, and make it work on non
pacea objects also.
##'
}
\examples{
\dontrun{
overlap(hake_recruitment, bi)
overlap(herring_spawning_biomass,
}
}
\author{
Andrew Edwards
}
