% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear-cache.R
\name{pacea_clear_cache}
\alias{pacea_clear_cache}
\title{pacea clear cache}
\usage{
pacea_clear_cache()
}
\description{
Delete all data files in cache folder.
}
\details{
This function will prompt the user on whether to delete all files in the cache directory. This cannot be undone and the user will have to download the files again.
}
\examples{
\dontrun{
bccm_avg0to40m_oxygen()
list.files(pacea_cache())
pacea_clear_cache()
}

}
