% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-biomass-herring.R
\name{plot.pacea_biomass_herring}
\alias{plot.pacea_biomass_herring}
\title{Plot a herring biomass time series object for one or all herring regions}
\usage{
\method{plot}{pacea_biomass_herring}(
  obj,
  region = NULL,
  x_lab = "Year",
  y_lab = attr(obj, "axis_name"),
  mar_all_regions = c(3, 3, 2, 0),
  oma_all_regions = c(2, 1, 1, 1),
  y_tick_by = 5,
  title = "full",
  ...
)
}
\arguments{
\item{obj}{a `pacea_biomass_herring` object, which is a tibble of
annual spawning biomass or total biomass estimates (TODO not yet
implemented) for each of the five herring Stock Assessment Regions.}

\item{region}{if `NULL` (the default) then plots a panel plot of biomass
for all five regions, else plots the specified region, which must be one
of HG, PRD, CC, SOG, or WCVI.}

\item{x_lab}{label for x-axis (default is "Year", only shown for bottom
panel when plotting all five regions)}

\item{y_lab}{label for y-axis (default is specified by `attr(obj, "axis_name")`.}

\item{mar_all_regions}{`mar` value for `par()`, may need tweaking for
five-panel plot; for single-panel just specific `mar`; see `?par`}

\item{oma_all_regions}{similar to `mar_all_regions`}

\item{y_tick_by}{increment for y-axis ticks; gets
overwritten in `add_tickmarks()` if this yields more than
`y_tick_max_number` tickmarks. If using [plot.pacea_biomass()] the default
of 1 gets automatically changed to 0.25 for `plot(hake_biomass)`. If
`y_axis_reverse` is TRUE then the negative of `y_tick_by` (which should
always be positive) is passed onto `add_tickmarks()`.}

\item{title}{if `NULL` then no figure title (using `main()`), if `full` (the
default) then spell out the Stock Assessment Region, and if `short` then
just use the acronym.}

\item{...}{further options passed onto `plot.pacea_biomass()` that can
also pass onto `plot.default()`}
}
\value{
plot of the herring biomass time series as joined up medians with
  shaded regions showing uncertainty, for either all regions or the
  specified region.
}
\description{
Adapted slightly from `plot.pacea_recruitment_herring()`. Plots in the
biomass style developed in `plot.pacea_biomass()`.
}
\examples{
\dontrun{ TODO

}
}
\author{
Andrew Edwards
}
