% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-buoy.R
\name{plot.pacea_buoy}
\alias{plot.pacea_buoy}
\title{Plot buoy SST data for one or multiple stations}
\usage{
\method{plot}{pacea_buoy}(
  obj,
  stn_id = "C46205",
  years = NULL,
  year_highlight = lubridate::year(lubridate::today())
)
}
\arguments{
\item{obj}{buoy SST data, of class `pacea_buoy_sst`}

\item{stn_id}{single station to plot, as given by
`stn_id` column in `buoy_metadata`. Default is station C46205 (West Dixon
Entrance, north of Haida Gwaii)}

\item{years}{vector of given years to plot. If left as NULL then plots all
available years.}

\item{year_highlight}{numeric of the year to highlight; defaults to current year.}
}
\value{
a ggplot object of the plot
}
\description{
Plot buoy SST data for either one station or multiple stations, with various
options, with a certain year (defaults to the current year), highlighted in
red.
}
\examples{
\dontrun{
plot(buoy_sst)
plot(buoy_sst, stn_id = "C46185")
}
}
\author{
Andrew Edwards
}
