% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-buoy-anomalies-list.R,
%   R/plot-pacea-buoy-anomalies-single.R
\name{plot.pacea_buoy_anomalies_list}
\alias{plot.pacea_buoy_anomalies_list}
\alias{plot_pacea_buoy_anomalies_single}
\title{Plot anomalies of the buoy sea-surface temperature data}
\usage{
\method{plot}{pacea_buoy_anomalies_list}(
  pacea_buoy_anomalies_list,
  stn_id_to_plot = NULL,
  months = NULL,
  main = NULL,
  xlab = "Year",
  ylab = "Buoy",
  use_stn_id_name = TRUE,
  sst_plot = "anomalies",
  count_breaks = NULL,
  return_results = FALSE,
  require_requested_months = NULL,
  scale_limits = NULL
)

plot_pacea_buoy_anomalies_single(
  pacea_buoy_anomalies_list,
  stn_id_to_plot,
  months,
  main,
  xlab,
  ylab,
  use_stn_id_name,
  sst_plot = "anomalies",
  count_breaks = NULL,
  return_results = FALSE,
  scale_limits = NULL
)
}
\arguments{
\item{pacea_buoy_anomalies_list}{object of class \code{pacea_buoy_anomalies_list}
obtained from running \code{calculate_anomalies()} on buoy data.}

\item{stn_id_to_plot}{single \code{stn_id} to plot}

\item{months}{numeric vector of months (1-12) to include. If not specified
then defaults to 4
(April), except when only one \code{stn_id_to_plot} when all months are plotted (unless specified). Can have more than one month, e.g. 6:9. For a 'winter' average,
say Nov-Mar, specify the months as \code{c(11, 12, 1, 2, 3)}. The winter average
anomaly will be calculated and named for the year in which January
falls. \code{months} must be consecutive (except for the overwintering \verb{12, 1} as
in the above example).}

\item{main}{title for the plot, if \code{NULL} then created automatically,
including detailing the months selected. May need to manually specify \code{main}
if many non-consecutive months are chosen (which seems unlikely).}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{use_stn_id_name}{logical, whether to use the name of the buoy
(e.g. \verb{Middle NOMAD}) or the \code{stn_id} (e.g. C46004). See \code{buoy_metadata} for
the names.}

\item{sst_plot}{character, one of:
\itemize{
\item \code{"anomalies"} (default) - plots SST anomalies
\item \code{"mean"} - plots mean SST values for each month
\item \code{"count"} - plots count of daily SST values used to calculate monthly mean
}}

\item{count_breaks}{numeric vector of break points for the count plot colour scale.
Only used when \code{sst_plot = "count"}. Default of \code{NULL} actually defaults to
\code{c(0, 10, 15, 20, 31) * length(months)} if \code{months} is specified, else it is
\code{c(0, 10, 15, 20, 31)}.}

\item{return_results}{logical, if \code{FALSE} (default) returns the plot object only.
If \code{TRUE}, prints the plot and returns a list with both \code{plot} and \code{results}.}

\item{require_requested_months}{numeric, number of months that must be available
out of the requested \code{months} to compute an average anomaly. If \code{NULL} (default),
defaults to \code{length(months)}. Only used when \code{length(months) > 1} and
\code{length(stn_id_to_plot) > 1}.}

\item{scale_limits}{numeric vector of length 2 specifying the limits for the colour scale.
If \code{NULL} (default), limits are calculated automatically based on the data. Only used
for non-count plots. If any plotted values are outside of \code{scale_limits}
then they are coloured grey.}
}
\value{
a ggplot object (when \code{return_results = FALSE}) or a list with \code{plot} and \code{results}
(when \code{return_results = TRUE})

a ggplot object
}
\description{
TODO Need to write this.
##'
}
\examples{
\dontrun{
# See the vignette for explanations and use of the options.
all_buoys_anomalies <- calculate_anomalies(buoy_sst,
                                           climatology_time = "month")
plot(all_buoys_anomalies)
}
\dontrun{
# See the vignette
}
}
\author{
Andrew Edwards
}
