% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-index.R
\name{plot.pacea_index}
\alias{plot.pacea_index}
\title{Plot a pacea climatic or oceanographic index}
\usage{
\method{plot}{pacea_index}(
  obj,
  value = "anomaly",
  xlab = "Date",
  ylab = attr(obj, "axis_name"),
  smooth_over_year = FALSE,
  type = "l",
  style = "red_blue_bar",
  y_tick_by = 0.25,
  y_tick_start = NULL,
  y_tick_end = NULL,
  x_tick_extra_years = 20,
  start_decade_ticks = lubridate::ymd("1800-01-01", truncated = 2),
  ...
)
}
\arguments{
\item{obj}{a `pacea_index` object, which is a time series.}

\item{value}{which column of `obj` to plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label, the default is an attribute of the `pacea_index`
object. TODO Note that this isn't automated yet to refer to anomaly or absolute
values, though for each time series there is probably only one choice}

\item{smooth_over_year}{logical to smooth monthly values over each calendar
year (as per Tetjana Ross' plots, see `?oni` for reference). Note that the
corresponding `date` is for 1st January of each year. TODO when
smoothed the red-blue figure isn't quite right (it isn't for monthly also,
but that's not as obvious). See Issue #15.}

\item{type}{usual argument for `plot()`}

\item{style}{what style of plot -- HERE TODO mention don't all work properly
but can tweak lwd
for "red_blue_bar" (default) for red bars above 0 and
blue bars below 0, "red_blue" for colouring red above 0 and
blue below (TODO needs splines or similar to smooth), TODO to implement:
"goa" for Gulf of Alaska Ecosystem Report style plots; "plain"
for just a line.}

\item{y_tick_by}{increment for y-axis ticks}

\item{y_tick_start}{where to start y-axis tickmarks, set automatically if not
specified (may need to occasionally specify)}

\item{y_tick_end}{where to end y-axis tickmars, as for `y_tick_start`}

\item{x_tick_extra_years}{number of extra years to expand around the range
of data for which to add annual tick marks (does not expand the axis)}

\item{start_decade_ticks}{where to start tickmarks for decades (defaults to
1800 as hard to automate)}

\item{...}{optional arguments passed onto `plot()`. Note that the x-axis is
constructed using a lubridate `date` object, so `xlim` needs to be a
`date` object (see example).}
}
\value{
plot of the time series to the current device (returns nothing)
}
\description{
Temporal plot for a pacea index (`pacea_index`) object, with options for
display style.
}
\examples{
\dontrun{
plot(oni)
plot(oni,
     xlim = c(lubridate::dmy(01011950),
     lubridate::dmy(01012040))) # to expand x-axis
plot(npi_monthly,
     value = "value")
}
}
\author{
Andrew Edwards
}
