% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-index.R
\name{plot.pacea_index}
\alias{plot.pacea_index}
\title{Plot a pacea climatic or oceanographic index}
\usage{
\method{plot}{pacea_index}(
  obj,
  value = "anomaly",
  xlab = "Date",
  ylab = attr(obj, "axis_name"),
  smooth_over_year = FALSE,
  type = "l",
  style = "red_blue_bar",
  y_tick_by = 0.25,
  y_tick_start = NULL,
  y_tick_end = NULL,
  y_tick_max_number = 200,
  x_tick_extra_years = 200,
  start_decade_ticks = lubridate::ymd("1800-01-01", truncated = 2),
  event_years = NULL,
  event_lub = NULL,
  event_pch = 20,
  event_cex = 3,
  event_col = "grey",
  ...
)
}
\arguments{
\item{obj}{a `pacea_index` object, which is a time series.}

\item{value}{which column of `obj` to plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label, the default is an attribute of the `pacea_index`
object.}

\item{smooth_over_year}{logical to smooth monthly values over each calendar
year (as per Tetjana Ross' plots, see `?oni` for reference). Note that the
corresponding `date` is for 1st January of each year.}

\item{type}{usual argument for `plot()`}

\item{style}{what style of plot:
\describe{
  \item{"red_blue_bar":}{for red bars above 0 and
  blue bars below 0, but may need to manually adjust the thickness `lwd` for
  width of bars to look good for your specific plot}
  \item{"red_blue":}{for a line with filled in colouring for red above 0 and
  blue below 0}
  \item{"goa":}{TODO (not implemented yet) for Gulf of Alaska Ecosystem
  Report style plots}
  \item{"plain":}{just a plain line}
}}

\item{y_tick_by}{increment for y-axis ticks (for [plot.pacea_biomass()] gets
overwritten in `add_tickmarks()` if this yields more than
`y_tick_max_number` tickmarks.}

\item{y_tick_start}{where to start y-axis tickmarks, set automatically if not
specified (may need to occasionally specify)}

\item{y_tick_end}{where to end y-axis tickmars, as for `y_tick_start`}

\item{y_tick_max_number}{maximum number of y tickmarks.}

\item{x_tick_extra_years}{number of extra years to expand around the range
of data for which to add annual tick marks (does not expand the axis); in
hindsight could have simplified this in `add_tickmarks()`, but just made
the default big here.}

\item{start_decade_ticks}{where to start tickmarks for decades (defaults to
1800 as hard to automate)}

\item{event_years}{years of the event occurring (e.g. sighting of a rare
shark) to add to the plot (sensible for rare-ish events, to quickly see if
they coincide with the values of the index);
points will be shown for 1st July (Oh Canada!) of the given year, and at
the value of the index to make it easy to quickly spot any
correlation. Assumes `event_years` values are unique.
For more control instead use `event_lub`.}

\item{event_lub}{dates of events as lubridate objects (only `event_years` or
`event_lub` can be specified). Assumes `event_lub` values are unique.}

\item{event_pch}{`pch` for events}

\item{event_cex}{`cex` for events}

\item{event_col}{`col` for events}

\item{...}{optional arguments passed onto `plot()`. Note that the x-axis is
constructed using a lubridate `date` object, so `xlim` needs to be a
`date` object (see example).}
}
\value{
plot of the time series to the current device (returns nothing)
}
\description{
Temporal plot for a pacea index (`pacea_index`) object, with options for
display style, and adding in times of other events (to ask questions such as
'does this rare event coincide with El Niño?'). See examples and vignette.
}
\examples{
\dontrun{
plot(oni)
plot(oni,
     xlim = c(lubridate::dmy(01011950),
     lubridate::dmy(01012040))) # to expand x-axis
plot(npi_monthly,
     value = "value")

# The overlaying of events idea was inspired from a lunchtime conversation regarding a
# rare sighting of a Bluntnose Sixgill Shark by divers in Port Alberni, that garnered
# [media attention](https://www.timescolonist.com/local-news/divers-encounter-deep-water-shark-in-alberni-inlet-7102038).
# A lunchtime conversation with Maxime Veilleux led to the question of whether catches
# f these sharks in BC waters are related to El Niño. Specifically, as the data from
# the International Pacific Halibut Commission annual longline survey are
# readily available in the [gfiphc](https://github.com/pbs-assess/gfiphc)
# package, we looked at years which caught a Bluntnose Sixgill Shark (at a
# standard station, outside of the Strait of Georgia).

library(gfiphc)
sp_set_counts <- iphc_get_calc_plot_full("bluntnose sixgill shark")   # Need
  access to DFO Groundfish database
bluntnose_caught_years <- unique(filter(sp_set_counts$set_counts, N_it > 0, standard == "Y")$year)
plot(oni,
     event_years = bluntnose_caught_years,
     xlim = c(lubridate::dmy(01011995), lubridate::dmy(01012024)),
     lwd = 2)
# This is not meant to be a definitive analysis, but a demonstration of the plotting options in pacea.
}
}
\author{
Andrew Edwards
}
