% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-oi.R
\name{plot.pacea_oi}
\alias{plot.pacea_oi}
\title{Plot OISST data}
\usage{
\method{plot}{pacea_oi}(x, weeks.plot, months.plot, years.plot, bc = TRUE, eez = TRUE, ...)
}
\arguments{
\item{x}{a OISST `pacea_oi` object, which is an `sf` object.}

\item{weeks.plot}{numeric vector to indicate which weeks to plot. Defaults to current week (or most recent) available.}

\item{months.plot}{character or numeric vector to indicate which months to plot (e.g. `c(1, 2)`, `c("April", "may")`, `c(1, "April")`). Defaults to current month (or most recent) available.}

\item{years.plot}{numeric vector to indicate which years to plot. Defaults to current year (or most recent) available.}

\item{bc}{logical. Should BC coastline layer be plotted?}

\item{eez}{logical. Should BC EEZ layer be plotted?}

\item{...}{other arguments to be passed on, but not currently used (`?ggplot`
says the same thing); this should remove a R-CMD-check warning.}
}
\value{
plot of the spatial data to the current device (returns nothing)
}
\description{
Plot for NOAA OISST sf objects using `ggplot()`. A quick visualization of data, specifying month(s) and year(s). For more options and configurable plots see vignette.
}
\examples{
\dontrun{
pdata <- oisst_7day
plot(pdata)
}
}
