% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-oianom.R
\name{plot.pacea_oianom}
\alias{plot.pacea_oianom}
\title{Plot anomaly of OISST spatiotemporal data layer}
\usage{
\method{plot}{pacea_oianom}(
  x,
  weeks.plot,
  months.plot,
  years.plot,
  clim.dat,
  bc = TRUE,
  eez = TRUE,
  ...
)
}
\arguments{
\item{x}{an OISST \code{pacea_oianom} object; output from using \code{calc_anom()} of \code{oisst_7day} or \code{oisst_month} data}

\item{weeks.plot}{weeks to plot. Defaults to current week (if available)}

\item{months.plot}{months to plot. Defaults to current month (if available)}

\item{years.plot}{years to plot. Defaults to current year (if available)}

\item{clim.dat}{climatology data, obtained from using \code{calc_clim()}. If used, contours of deviations from climatology mean will be plotted}

\item{bc}{logical. Should BC coastline layer be plotted? Can only be plotted with one plot layer.}

\item{eez}{logical. Should BC EEZ layer be plotted? Can only be plotted with one plot layer.}

\item{...}{other arguments to be passed on, but not currently used (\code{?ggplot}
says the same thing); this should remove a R-CMD-check warning.}
}
\value{
plot of the spatial data to the current device (returns nothing)
}
\description{
Gets called from \code{pacea_oisst_anomalies_list()}, which uses the newer format
of climatology and anomalies TODO. Ensure consistent.
}
\examples{
\dontrun{
anom_data <- calc_anom(oisst_7day)
plot(anom_data)
}
}
