% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-oisst-anomalies-list.R
\name{plot.pacea_oisst_anomalies_list}
\alias{plot.pacea_oisst_anomalies_list}
\title{Plot anomalies of the OISST sea-surface temperature values, as either a
single or series of spatial plots, or a heatmap style.}
\usage{
\method{plot}{pacea_oisst_anomalies_list}(pacea_oisst_anomalies_list, heatmap_style = FALSE, ...)
}
\arguments{
\item{pacea_oisst_anomalies_list}{object of class \code{pacea_oisst_anomalies_list}
obtained from running \code{caclulate_anomalies()} on oisst values.}

\item{heatmap_style}{logical, if \code{FALSE} (default) then do a spatial plot of
the absolute anomalies with one map for each month and year, with default
(\code{months = NULL} being the latest month of data. See \code{\link[=plot.pacea_oianom]{plot.pacea_oianom()}} for arguments.
If \code{heatmap_style = TRUE} then do a heatmap style figure showing the
absolute anomalies for each month averaged over the designated area. Has
January at the top and December at the bottom of the y-axis, with years
along the x-axis. See \code{\link[=plot_pacea_oisst_anomalies_heatmap_style]{plot_pacea_oisst_anomalies_heatmap_style()}} for
arguments.}

\item{...}{arguments to be passed onto
\code{\link[=plot_pacea_oisst_anomalies_heatmap_style]{plot_pacea_oisst_anomalies_heatmap_style()}} or \code{\link[=plot.pacea_oianom]{plot.pacea_oianom()}}.}
}
\description{
See \code{heatmap_style} arguments for the resulting type of plot.
}
\examples{
\dontrun{
# See the vignette for explanations and use of the options.
oisst_anomalies <- calculate_anomalies(oisst_month)
plot(oisst_anomalies)         # spatial map of anomalies for most recent much
plot(oisst_anomalies, heatmap_style = TRUE)   # mean anomaly for each month
                                   # averaged over the whole spatial domain
}
}
\author{
Andrew Edwards and Andrea Hilborn
}
