% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-recruitment.R
\name{plot.pacea_recruitment}
\alias{plot.pacea_recruitment}
\title{Plot a pacea recruitment time series object (currently assumes annual values)}
\usage{
\method{plot}{pacea_recruitment}(
  obj,
  value = "median",
  xlab = "Year",
  ylab = attr(obj, "axis_name"),
  y_tick = 0.25,
  x_tick_extra_years = 20,
  style = "no_uncertainty",
  uncertainty_bar_col = "blue",
  y_max = NULL,
  ...
)
}
\arguments{
\item{obj}{a `pacea_recruitment` object, which is a time series. Function
will run on other objects (not give an error) but is not tested on those.}

\item{value}{which column of `obj` to plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label, the default is an attribute of the `pacea_index`
object. TODO Note that this isn't automated yet to refer to anomaly or absolute
values, though for each time series there is probably only one choice}

\item{y_tick}{increment for y-axis ticks}

\item{x_tick_extra_years}{number of extra years to expand around the range
of data for which to add annual tick marks}

\item{style}{what style of plot -- HERE TODO mention don't all work properly
but can tweak lwd
for "red_blue_bar" (default) for red bars above 0 and
blue bars below 0, "red_blue" for colouring red above 0 and
blue below (TODO needs splines or similar to smooth), TODO to implement:
"goa" for Gulf of Alaska Ecosystem Report style plots; "plain"
for just a line.}

\item{uncertainty_bar_col}{colour for uncertainty bars for certain types of
plot (e.g. estimated fish recruitment)}

\item{y_max}{maximum y value for certain types of plot (use this if you get
an error when specifying `ylim`)}

\item{...}{optional arguments passed onto `plot()`. Note that the x-axis is
constructed using a lubridate `date` object, so `xlim` needs to be a
`date` object (see example).}
}
\value{
plot of the time series as median with bars showing uncertainty (if
  `low` and `high` are columns of `obj) to the current device; returns nothing.
}
\description{
Temporal plot for a pacea recruitment time series (of class
`pacea_recruitment`) object. The `style` option here (unlike for
`plot.pacea_index()` defaults to `no_uncertainty` and gets changed to
`uncertainty` if `low` and `high` are columns of `obj`.
}
\examples{
\dontrun{
plot(hake_recruitment)
plot(hake_recruitment,
     xlim = c(lubridate::dmy(01011950),
              lubridate::dmy(01012040))) # to expand x-axis
}
}
\author{
Andrew Edwards
}
