% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-harbour-seals-gam.R,
%   R/plot-pacea-recruitment.R
\name{plot.pacea_recruitment}
\alias{plot.pacea_recruitment}
\title{Plot a pacea recruitment time series object (currently assumes annual values)}
\usage{
\method{plot}{pacea_recruitment}(
  obj,
  value = "median",
  xlab = "Year",
  ylab = attr(obj, "axis_name"),
  y_tick_by = 1,
  x_tick_extra_years = 20,
  start_decade_ticks = lubridate::ymd("1800-01-01", truncated = 2),
  style = "no_uncertainty",
  uncertainty_bar_col = "blue",
  y_max = NULL,
  add_line_at_1 = FALSE,
  add_line_at_1_col = "darkgreen",
  add_line_at_1_lty = 2,
  ...
)

\method{plot}{pacea_recruitment}(
  obj,
  value = "median",
  xlab = "Year",
  ylab = attr(obj, "axis_name"),
  y_tick_by = 1,
  x_tick_extra_years = 20,
  start_decade_ticks = lubridate::ymd("1800-01-01", truncated = 2),
  style = "no_uncertainty",
  uncertainty_bar_col = "blue",
  y_max = NULL,
  add_line_at_1 = FALSE,
  add_line_at_1_col = "darkgreen",
  add_line_at_1_lty = 2,
  ...
)
}
\arguments{
\item{obj}{a `pacea_recruitment` object, which is a time series. Function
will run on other objects (not give an error) but is not tested on those.}

\item{value}{the column to plot if no uncertainties, or what to plot as dots
if showing uncertainties (likely always `median`)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label, the default is an attribute of the `pacea_index`
object.}

\item{y_tick_by}{increment for y-axis ticks}

\item{x_tick_extra_years}{number of extra years to expand around the range
of data for which to add annual tick marks (does not expand the axis); in
hindsight could have simplified this in `add_tickmarks()`, but just made
the default big here.}

\item{start_decade_ticks}{where to start tickmarks for decades (defaults to
1800 as hard to automate)}

\item{style}{`no_uncertainty` for plain time series without uncertainty,
gets overridden to have uncertainty bars if `low` and `high` are columns
of `obj`}

\item{uncertainty_bar_col}{colour for uncertainty bars for certain types of
plot (e.g. estimated fish recruitment)}

\item{y_max}{maximum y value for certain types of plot (use this if you get
an error when specifying `ylim`)}

\item{add_line_at_1}{whether to add a horizontal line at 1 (only sensible for scaled recruitments)}

\item{add_line_at_1_col}{colour for line at 1}

\item{add_line_at_1_lty}{line type of line at 1}

\item{...}{further options passed onto `plot.default()`}
}
\value{
plot of the time series as median with bars showing uncertainty (if
  `low` and `high` are columns of `obj) to the current device; returns nothing.

plot of the time series as median with bars showing uncertainty (if
  `low` and `high` are columns of `obj) to the current device; returns nothing.
}
\description{
Temporal plot for a pacea recruitment time series (of class
`pacea_recruitment`) object. The `style` option here (unlike for
`plot.pacea_index()` defaults to `no_uncertainty` and gets changed to
`uncertainty` if `low` and `high` are columns of `obj`.

Temporal plot for a pacea recruitment time series (of class
`pacea_recruitment`) object. The `style` option here (unlike for
`plot.pacea_index()` defaults to `no_uncertainty` and gets changed to
`uncertainty` if `low` and `high` are columns of `obj`.
}
\examples{
\dontrun{
plot(hake_recruitment)
plot(hake_recruitment,
     xlim = c(lubridate::dmy(01011950),
              lubridate::dmy(01012040))) # to expand x-axis
plot(hake_recruitment_over_2010)  # automatically changes style of plot
                                  #  if 'over' is in the object name
}
\dontrun{
plot(hake_recruitment)
plot(hake_recruitment,
     xlim = c(lubridate::dmy(01011950),
              lubridate::dmy(01012040))) # to expand x-axis
plot(hake_recruitment_over_2010)  # automatically changes style of plot
                                  #  if 'over' is in the object name
}
}
\author{
Andrew Edwards
}
