% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-st.R
\name{plot.pacea_st}
\alias{plot.pacea_st}
\title{Plot a pacea spatiotemporal data layer}
\usage{
\method{plot}{pacea_st}(
  obj,
  months.plot = c("April"),
  years.plot = c(2018),
  bc = TRUE,
  eez = TRUE
)
}
\arguments{
\item{obj}{a BCCM ROMS `pacea_st` object, which is an `sf` object}

\item{months.plot}{character or numeric vector to indicate which months to include (e.g. `c(1, 2)`, `c("April", "may")`, `c(1, "April")`)}

\item{years.plot}{vector of years to include, from 1993 to 2019}

\item{bc}{logical. Should BC coastline layer be plotted?}

\item{eez}{logical. Should BC EEZ layer be plotted?}

\item{...}{optional arguments passed on to `plot.sf()`}
}
\value{
plot of the spatial data to the current device (returns nothing)
}
\description{
Plot for BCCM ROMS sf objects using `ggplot()`. A quick visualization of data, specifying month(s) and year(s). For more options and configurable plots see vignette.
}
\examples{
\dontrun{
pdata <- bccm_surface_temperature()
plot(pdata)
}
}
