% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-st.R
\name{plot.pacea_st}
\alias{plot.pacea_st}
\title{Plot a pacea spatiotemporal data layer}
\usage{
\method{plot}{pacea_st}(
  x,
  months.plot = c("April"),
  years.plot = c(2018),
  bc = TRUE,
  eez = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{pacea_st} object, which is an \code{sf} object}

\item{months.plot}{character or numeric vector to indicate which months to include (e.g. \code{c(1, 2)}, \code{c("April", "may")}, \code{c(1, "April")})}

\item{years.plot}{vector of years to include, from 1993 to 2019}

\item{bc}{logical. Should BC coastline layer be plotted?}

\item{eez}{logical. Should BC EEZ layer be plotted?}

\item{...}{other arguments to be passed on, but not currently used (\code{?ggplot}
says the same thing); this should remove a R-CMD-check warning.}
}
\value{
plot of the spatial data to the current device (returns nothing)
}
\description{
Plot for \code{pacea_st} classes of objects using \code{ggplot()}. Objects can be BCCM
outputs, OISST values (TODO no, but should work for the grid26 mapped values), or HOTSSEA outputs.
Gives a quick visualization of data, specifying month(s) and year(s). For
more options and configurable plots see vignette.
}
\examples{
\dontrun{
pdata <- bccm_surface_temperature()
plot(pdata)
}
}
