% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-zooplankton.R
\name{plot.pacea_zooplankton}
\alias{plot.pacea_zooplankton}
\title{Plot the one of the zooplankton anomalies.}
\usage{
\method{plot}{pacea_zooplankton}(
  obj,
  species_group = "total_biomass",
  xlab = "Date",
  ylab = NULL,
  lwd = 15,
  ...
)
}
\arguments{
\item{obj}{a `pacea_zooplankton` object, which is a tibble with columns
giving `year`, `num_samples` (number of samples), `volume_filtered`,
`total_biomass`, and then 24 species groups of zooplankton (see `?zooplankton`).}

\item{species_group}{which column of `obj` to plot, either `total_biomass`
(the default) or one of the 24 species groups described in `?zooplankton`.}

\item{xlab}{x-axis label, defaults to `Date`}

\item{ylab}{y-axis label, if `NULL` (the default` then automatically selects
the `axis_name` corresponding to `species_group` in the object
`zooplankton_axis_names`. User can also just provide their own label.}

\item{lwd}{line thickness to plot the anomalies, defaults to 15 since these
are annual; likely needs tweaking depending on the size of the final figure.}

\item{...}{optional arguments passed onto `plot.pacea_index()` and
`plot()`. Note that the x-axis is constructed using a lubridate `date`
object, so `xlim` needs to be a  `date` object (see `plot.pacea_index()` example).}
}
\value{
plot of the time series to the current device (returns nothing)
}
\description{
Temporal plot for one of the columns of a pacea zooplankton
(`pacea_zooplankton`) object, as defined in `?zooplankton`. The specified
`species_group` gets passed onto `plot.pacea_index()` as the anomaly to be
plotted, as do any options specified here for display style, adding in
times of other events, etc. See examples and vignette, as well as the
climatic and oceanographic indices vignette.
}
\examples{
\dontrun{
plot(oni)
plot(oni,
     xlim = c(lubridate::dmy(01011950),
     lubridate::dmy(01012040))) # to expand x-axis
plot(npi_monthly,
     value = "value")
plot(oni,
     event_years = bluntnose_caught_years,
     xlim = c(lubridate::dmy(01011995), lubridate::dmy(01012024)),
     lwd = 2)
}
}
\author{
Andrew Edwards
}
