% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-buoy-sst.R
\name{plot_buoy_sst_single}
\alias{plot_buoy_sst_single}
\title{Plot buoy SST data for a single station; called by `plot.pacea_buoy()`}
\usage{
plot_buoy_sst_single(obj, year_highlight, title = NULL)
}
\arguments{
\item{obj}{TODO}

\item{title}{TODO}
}
\description{
<desc>
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
