% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-oisst-anomalies-heatmap-style.R
\name{plot_pacea_oisst_anomalies_heatmap_style}
\alias{plot_pacea_oisst_anomalies_heatmap_style}
\title{Plot anomalies of OISST data averaged over a defined area, shown with
months on the y-axis in a heatmap style plot.}
\usage{
plot_pacea_oisst_anomalies_heatmap_style(
  dat,
  area = NULL,
  area_name = NULL,
  main = NULL,
  months = NULL,
  xlab = "Year",
  ylab = "Month",
  return_results = FALSE,
  scale_limits = NULL
)
}
\arguments{
\item{dat}{\code{pacea_oisst_anomalies_list} class object from running
\code{calculate_anomalies()} on an object of class \code{pacea_oi} (the \code{sst_month}
data).}

\item{area}{sf object of the area to average the SST over to get a single
value for each month (uses any cell of the OISST data that intersects with \code{area}.}

\item{area_name}{character string of the name of the area to use in the
title of the plot. If \code{NULL} then no area is specified (but we recommend
specifying one for clarity).}

\item{main}{title for the plot, if left as \code{NULL} then
created automatically, including using \code{area_name}. May need to manually
specify \code{main} as it is hard to fully automate.}

\item{months}{numeric vector of months (1-12) to show; default (\code{NULL})
shows all 12 months.}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{return_results}{logical, if \code{FALSE} (default) returns (plots) the ggplot object only.
If \code{TRUE}, prints the plot and returns a list with both \code{plot} and \code{results}.}

\item{scale_limits}{numeric vector of length 2 specifying the limits for the colour scale.
If \code{NULL} (default), limits are calculated automatically to give a symmetric
scale based on the anomalies. If any plotted values are outside of \code{scale_limits}
then they are coloured grey.}
}
\value{
a ggplot object (when \code{return_results = FALSE}) or a list with \code{plot} and \code{results}
(when \code{return_results = TRUE})
}
\description{
Gets called from \code{plot.pacea_oisst_anomalies_list()} if \code{heatmap_style == TRUE}.
}
\examples{
\dontrun{
# See the vignette for explanations and use of the options.
oisst_anomalies <- calculate_anomalies(oisst_month)
plot(oisst_anomalies)         # spatial map of anomalies for most recent much
plot(oisst_anomalies, heatmap_style = TRUE)   # mean anomaly for each month
                                   # averaged over the whole spatial domain
}
}
\author{
Andrew Edwards and Andrea Hilborn
}
