% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pacea-index.R
\name{plot_red_blue}
\alias{plot_red_blue}
\title{Plot the red/blue style of anomaly plot, linearly interpolating to make
smooth (needed when crossing x-axis); internal function called from `plot-pacea_index()`.}
\usage{
plot_red_blue(obj_lub, value, xlab, ylab, type, ...)
}
\arguments{
\item{obj_lub}{obj a `pacea_index` object, which is a time series, with a date
column that is the lubridate `date` class.}

\item{value}{which column of `obj` to plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label, the default is an attribute of the `pacea_index`
object.}

\item{type}{usual argument for `plot()`}

\item{...}{optional arguments passed onto `plot()`. Note that the x-axis is
  constructed using a lubridate `date` object, so `xlim` needs to be a
  `date` object (see example).
##' @return plot of the time series to the current device (returns nothing)}
}
\value{
plot of time series
}
\description{
Adapted from
https://stackoverflow.com/questions/74902499/shading-below-line-graph-in-r/74903305#74903305
and interpolated.
}
\examples{
\dontrun{
# see plot.pacea_index()
}
}
\author{
Andrew Edwards
}
