% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-with-uncertainty.R
\name{plot_with_uncertainty}
\alias{plot_with_uncertainty}
\title{Plot medians with uncertainty bars, such as for stock assessment
recruitment; internal function called from `plot.pacea_recruitment()`.}
\usage{
plot_with_uncertainty(
  obj_lub,
  value,
  xlab,
  ylab,
  y_tick,
  x_tick_extra_years,
  uncertainty_bar_col,
  y_max,
  add_line_at_1,
  add_line_at_1_col,
  add_line_at_1_lty,
  ...
)
}
\arguments{
\item{obj_lub}{obj a `pacea_recruitment` object, which is a time series, with a date
column that is the lubridate `date` class. Can be absolute or relative
recruitment (scaled by a particular year's recruitment or by assumed
unfished equilibrium recruitment).}

\item{value}{the column to plot if no uncertainties, or what to plot as dots
if showing uncertainties (likely always `median`)}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis (default is automatic from attribute of `obj`)}

\item{x_tick_extra_years}{number of extra years to expand around the range
of data for which to add annual tick marks}

\item{uncertainty_bar_col}{colour for uncertainty bars for certain types of
plot (e.g. estimated fish recruitment)}

\item{y_max}{maximum y value for certain types of plot (use this if you get
an error when specifying `ylim`)}

\item{add_line_at_1}{whether to add a horizontal line at 1 (only sensible for scaled recruitments)}

\item{add_line_at_1_col}{colour for line at 1}

\item{add_line_at_1_lty}{line type of line at 1}

\item{...}{further options passed onto `plot.default()`}
}
\value{
plot of time series with median as circle and bars for uncertainty.
}
\description{
Adapted from `make.mcmc.recruitment.plot()` from Pacific Hake
assessment.
}
\examples{
\dontrun{
# see plot.pacea_recruitment()
}
}
\author{
Andrew Edwards
}
