% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts-get-data.R
\name{roms_bottom_oxygen}
\alias{roms_bottom_oxygen}
\alias{roms_bottom_salinity}
\alias{roms_bottom_temperature}
\alias{roms_surface_oxygen}
\alias{roms_surface_salinity}
\alias{roms_surface_temperature}
\title{ROMS environmental variable data}
\format{
A simple features dataframe.
}
\usage{
roms_bottom_oxygen(ask = interactive())

roms_bottom_salinity(ask = interactive())

roms_bottom_temperature(ask = interactive())

roms_surface_oxygen(ask = interactive())

roms_surface_salinity(ask = interactive())

roms_surface_temperature(ask = interactive())
}
\arguments{
\item{ask}{Logical. Should the user be asked before downloading the data to local cache? Defaults to the value of interactive().}
}
\value{
`sf` data object requested.
}
\description{
Regional Ocean Modelling System (ROMS) data for the BC Pacific region.

ROMS data interpolated to a 2km x 2km inshore grid and a 6km x 6km offshore grid. 
Data are monthly averages from 1993-2019. Each column represents a distinct 'year_month' combination (e.g. 2010_4).

\code{roms_bottom_oxygen()} Bottom oxygen measured as UNITS [02]\cr
\code{roms_bottom_salinity()} Bottom salinity in UNITS!\cr
\code{roms_bottom_temperature()} Bottom temperature in oC\cr
\code{roms_surface_oxygen()} Surface oxygen measured as UNITS [02]\cr
\code{roms_surface_salinity()} Surface salinity in UNITS!\cr
\code{roms_surface_temperature()} Surface temperature in oC
}
\examples{
\dontrun{
my_data <- roms_bottom_oxygen()
}
}
