% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-months.R
\name{summarise_months}
\alias{summarise_months}
\title{Summarize a vector of months as a formatted string}
\usage{
summarise_months(months, just_first_to_last = FALSE)
}
\arguments{
\item{months}{numeric vector of month numbers (1-12)}

\item{just_first_to_last}{force function to return just the first and last
months in of \code{months}; generally wouldn't have gaps anyway}
}
\value{
character string summarizing the months
}
\description{
Summarize a vector of months as a formatted string
}
\examples{
summarise_months(4)             # "Apr"
summarise_months(c(4, 5))       # "Apr" and "May"
summarise_months(4:6)           # "Apr to Jun"
summarise_months(c(12, 1:3))    # "Dec, Jan, Feb, and Mar"
summarise_months(c(12, 1:3),
                 just_first_to_last = TRUE)
                                # "Dec to Mar"
Generally thinking we would have consecutive months, but can still spell
them out even if not:
summarise_months(c(12, 4:8))    #  "Dec, Apr, May, Jun, Jul, and Aug"

}
