% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-months.R
\name{summarise_months}
\alias{summarise_months}
\title{Summarize a vector of months as a formatted string}
\usage{
summarise_months(months)
}
\arguments{
\item{months}{numeric vector of month numbers (1-12)}
}
\value{
character string summarizing the months
}
\description{
Summarize a vector of months as a formatted string
}
\examples{
summarise_months(4)
summarise_months(c(4, 5))
summarise_months(c(4, 5, 6, 7))
summarise_months(c(4, 5, 7))

}
